==> RawCodes/step7 confounder.Rmd <==
---
title: "step7 confound"
author: "Ziyan Zhu"
date: "3/23/2021"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
source("step2_data_wrangle.R")
```

# Select varible of interests

```{r}
################### school reopen dates ################### 
district_policies <- OH_K12 %>%
  distinct(county,county_enroll,leaid,district_enroll,schooltemporaryshutdown,opendategrouped,teachingmethod)

table(district_policies[,c('teachingmethod','schooltemporaryshutdown')])
#table(district_policies[,c('county','opendategrouped')])
```

Most of the time:

Hybrid schools -- Never closed;

On Premises schools -- Never closed;

Online Only schools -- Never closed.

```{r}
district_policies%>%
  group_by(county,county_enroll,opendategrouped)%>%
  summarise(n_opendate = sum(district_enroll))%>%
  mutate(prop_opendate = round(n_opendate/county_enroll,2))%>%
  arrange(-prop_opendate)
```

# Calculate the proportion and generate date brackets

```{r}
prop_opendate <- district_policies%>%
  filter(!schooltemporaryshutdown %in% c('Closed indefinitely','Pending','Unknown'))%>%
  group_by(county,county_enroll,opendategrouped)%>%
  summarise(n_opendate = sum(district_enroll))%>% # number of students under certain date for each county
  mutate(prop_opendate = round(n_opendate/county_enroll,2))%>% # proportion
  group_by(county)%>%
  filter(prop_opendate>0.6)%>% # filter large proportions of students with same reopen dates #can be replaced with # slice(which.max(prop_opendate))#
  mutate(threeweeks_before = opendategrouped-21,sixweeks_later = as.Date('2020-12-15')+42)%>%
  select(-n_opendate)
```


# if you keep the filter there will be just 21 counties in the opendate_cases

```{r}
opendate_cases <- cases%>%
  inner_join(high_prop,by=c('COUNTY'='county'))%>%
  group_by(COUNTY)%>%
  filter(DATE>=threeweeks_before & DATE<=sixweeks_later)%>%
  group_by(COUNTY)%>%
  mutate(window_id = case_when(DATE<opendategrouped~"3 weeks before reopen",
    DATE>=opendategrouped & DATE<'2020-12-15' ~'fall semester',
    DATE>='2020-12-15' & DATE< as.Date('2020-12-15')+21 ~ '3 weeks after end',
    DATE>=as.Date('2020-12-15')+21 ~ '6 weeks after end'
  ))%>%
  select(-STATE,-STUSAB,-ST_LAT,-ST_LONG,-STATEFP,-GNISID,-UID,-CODE3)%>%
  mutate(death_prop = round(CUMDEATHS/POPULATION,5),window_id = as.factor(window_id))%>%
  left_join(wide_teaching_enroll,by=c('COUNTY'='county','county_enroll'))


# select the start date and end date data for each window of time
  
start_end_window <- opendate_cases%>%
  group_by(COUNTY,window_id)%>%
  arrange(DATE)%>%
  filter(row_number()==1 | row_number()==n())%>%
  arrange(COUNTY,DATE)%>%
  group_by(COUNTY,window_id)%>%
  mutate(y_labels = ifelse(row_number()==1,'y0','y1'))%>%
  dcast(COUNTY+POPULATION+county_enroll+opendategrouped+threeweeks_before+sixweeks_later+Online_Only+Hybrid+On_Premises+window_id~y_labels,value.var = 'CUMDEATHS')

```

```{r}
opendate_cases%>%
  ggplot(aes(x=DATE,y=death_prop,group=COUNTY,color=window_id))+geom_point()+facet_wrap(~major_teaching)+theme(legend.position = 'bottom')

opendate_cases%>%
  filter(major_teaching=='Hybrid')%>%
  ggplot(aes(x=DATE,y=death_prop,group=COUNTY,color=window_id))+geom_line(size=2)+facet_wrap(~major_teaching+COUNTY)+theme(legend.position = 'bottom')+geom_line()
```



--- down here is just for experiment -- split the fall semester by 21days window

```{r}
long_21d_y <- counties21%>%
  filter(DATE >= threeweeks_before,DATE <= sixweeks_later)%>%
  group_by(COUNTY)%>%
  filter(DATE == opendategrouped-21| DATE == opendategrouped|DATE == opendategrouped+21|DATE == opendategrouped+42|DATE == opendategrouped+63|DATE == opendategrouped+84|DATE == opendategrouped+105)

wide_21d_y <- long_21d_y%>%
  dcast(COUNTY+POPULATION+county_enroll+opendategrouped+threeweeks_before+Online_Only+Hybrid+On_Premises~DATE,value.var = "CUMDEATHS")

```


```{r fig.height=6, fig.width=6}
OH_K12%>%
  distinct(county,county_enroll,district_enroll,leaid,date,opendategrouped)%>%
ggplot(aes(x=as.Date(date),y=as.Date(opendategrouped)))+geom_point(aes(size = district_enroll),alpha=0.5,color = 'coral')+labs(x='School Year Start Date',y='School Reopen Date Grouped by District')+theme_minimal()+scale_x_date(date_labels = "%m-%d",limits = c(as.Date('2020-07-05'),as.Date('2020-12-31')),date_breaks = "3 weeks")+scale_y_date(date_labels = "%m-%d",limits = c(as.Date('2020-07-05'),as.Date('2020-12-31')),date_breaks = "3 weeks")
```


```{r}
open_reopen_enroll <- OH_K12%>%
  distinct(county,county_enroll,district_enroll,leaid,date)%>%
  group_by(county,date)%>%
  summarise(date_enroll = sum(district_enroll))%>%
  left_join(OH_K12%>%
  distinct(county,county_enroll,district_enroll,leaid,opendategrouped)%>%
  group_by(county,opendategrouped)%>%
  summarise(reopen_enroll = sum(district_enroll)), by = 'county' )
```
