==> RawCodes/step8_pre_graphs.Rmd <==
---
title: "Presentation Graphs"
author: "Cheyenne, Ziyan"
date: "4/4/2021"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r message=FALSE, warning=FALSE}
source("step2_data_wrangle.R")
```

# Case comparison by age

```{r}
library(readxl)
cases_by_age <- read_excel("OhiobyAge.xlsx")
```


```{r}
library(tidyverse)
library(lubridate)
require(scales)


rolling_age_cases <- cases_by_age %>%
  mutate(youth_prop_roll = zoo::rollmean(`00_19/total(%)`, k = 7, fill = NA),
         all_roll = zoo::rollmean(`00_80+`, k = 7, fill = NA))

colors <- c("Total Daily Cases" = "black",
            "0-19 Age / Total Cases (%)" = "gray")

coeff <- 200

ggplot(rolling_age_cases, aes(x=Date)) + 
  geom_line( aes(y=youth_prop_roll,
                 color = "0-19 Age / Total Cases (%)"), 
             na.rm = T)+
  geom_line( aes(y=all_roll/coeff,
                 color = "Total Daily Cases"), 
             na.rm = T) +
  scale_y_continuous(
    # Features of the first axis
    name = "% Daily Cases for 0-19 Year Age Group",
    labels = function(x){paste0(x, "%")},
    # Add a second axis and specify its features
    sec.axis = sec_axis(~.*coeff, name="New Daily Cases for Entire Population",
                        label=comma)
  ) +
  geom_rect(data=rolling_age_cases[1,],
            aes(xmin=as.POSIXct ("2020/08/18"), xmax=as.POSIXct ("2020/12/12"),
                ymin=-Inf,ymax=Inf),
            color = NA,alpha=0.2, show.legend = F, fill = "orange") + 
  geom_vline(xintercept = as.POSIXct ("2020/08/18") + days(7),lty = 2)+ 
  xlim(c(as.POSIXct ("2020/08/01"),as.POSIXct ("2021/01/01"))) + 
  labs(title = "Total Cases and Percent of Cases in 0-19 Year Age Group",
       subtitle = "Yellow Area represents the fall semester",
       caption = "Smoothed using a 7 day moving average",
       color = "")+
  scale_color_manual(values = colors)+ 
  theme_bw() + 
  theme(legend.position='bottom')

#ggsave("state_level_age.jpg",width = 7, height = 5)
```

```{r}
cases_by_age_long <- cases_by_age %>%
  gather(age_group, percent_cases, 
         `00_19/total(%)`:`80+/total(%)`,
         factor_key=TRUE) %>%
  group_by(age_group) %>%
  mutate(roll_percent_cases= zoo::rollmean(percent_cases, k = 7, fill = NA))

ggplot(cases_by_age_long,aes(x = Date, y = roll_percent_cases, color = age_group)) + 
  geom_line(na.rm = T) +
  scale_y_continuous(labels = function(x){paste0(x, "%")})+
  geom_rect(data=cases_by_age_long[1,],
            aes(xmin=as.POSIXct ("2020/08/18"), xmax=as.POSIXct ("2020/12/12"),
                ymin=-Inf,ymax=Inf),
            color = NA,alpha=0.2, show.legend = F, fill = "orange")  + xlim(c(as.POSIXct ("2020/08/01"),as.POSIXct ("2021/01/01"))) + 
  theme_bw() + 
  labs(title = "Percent of Cases by Age Group",
       subtitle = "Yellow Area represents the fall semester",
       y = "% Daily Cases",
       caption = "Smoothed using a 7 day moving average",
       color = "Age Group")+ theme_bw()+
  theme(legend.position="")+
  theme(axis.title = element_text(size=15),axis.text = element_text(size=15),title = element_text(size=15))+
  theme(axis.line = element_line(colour = "black"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_blank(),
    panel.background = element_blank()) 

#ggsave("prop_age.jpg", width = 7,height = 5)
```



## 
```{r}
# Plot
options(scipen=10000)


state_open_teaching_enroll%>%
  filter(teachingmethod %in% c("On Premises","Hybrid"))%>%
  ggplot(aes(y=opendate_teaching_county_enroll/1000, x=opendate,  fill=teachingmethod)) +
  geom_bar(position = 'stack', stat = 'identity',width = 3) +
  geom_rect(data = state_open_teaching_enroll[1,],
            aes(xmin = as.Date('2020/08/18'), 
                xmax = as.Date('2020/12/15'), ymin=-Inf, ymax=Inf), 
            color = NA, alpha = 0.2, show.legend = F, fill = 'orange') +
  labs(x='School Reopen Date',y = 'County Student Enrollment / 1000',title = 'Number of Student Enrollment on Different School Reopen Dates \n(On Premises & Hybrid)',col="",fill="Majority Teaching Method")+ xlim(as.Date('2020/08/01'),as.Date('2021/01/01'))+
  scale_fill_manual(values = c("#00BA38", "#F8766D")) +
  theme_bw()+theme(legend.position= c(0.77, 0.8),legend.title = element_text(size=13),legend.text = element_text(size=13),axis.title = element_text(size=15),axis.text = element_text(size=15),legend.background = element_rect(fill = alpha("orange",0.0)),legend.key.size = unit(1.4,"lines"),title = element_text(size=12.9))+
  theme(axis.line = element_line(colour = "black"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_blank(),
    panel.background = element_blank()) 


state_open_teaching_enroll%>%
  filter(teachingmethod %in% c("On Premises","Hybrid"))%>%
  ggplot(aes(y=opendate_teaching_state_prop, x=opendate,  fill=teachingmethod)) +
  geom_bar(position = 'stack', stat = 'identity',width = 3)+ylim(0,1)+
  geom_rect(data = state_open_teaching_enroll[1,],
            aes(xmin = as.Date('2020/08/18'), 
                xmax = as.Date('2020/12/15'), ymin=-Inf, ymax=Inf), 
            color = NA, alpha = 0.2, show.legend = F, fill = 'orange') +
  labs(x='School Reopen Date',y = 'County Student Proportion',title = 'Percent of Students on Different School Reopen Dates \n(teaching with in-person components)',col="",fill="Majority Teaching Method")+ xlim(as.Date('2020/08/01'),as.Date('2021/01/01'))+
  scale_fill_manual(values = c("#00BA38", "#F8766D")) +
  theme_bw()+theme(legend.position= c(0.77, 0.8),legend.title = element_text(size=13),legend.text = element_text(size=13),axis.title = element_text(size=15),axis.text = element_text(size=15),legend.background = element_rect(fill = alpha("orange",0.0)),legend.key.size = unit(1.4,"lines"),title = element_text(size=12.9))+
  theme(axis.line = element_line(colour = "black"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_blank(),
    panel.background = element_blank()) 

```

```{r}
county_policy_wide$major_teaching <- factor(county_policy_wide$major_teaching,levels = c("On Premises",
                                                       "Hybrid",
                                                      "Online Only"))

# see when the intesection happens
date.intercept <- as.Date("2020-11-24")


# add 95% confidence bans
confidence_level <- .95
z_cl <- qnorm(confidence_level)

# case_policy_wide
case_policy_wide <- cases %>%
  left_join(county_policy_wide[,c("county","major_teaching","Online_Only","Hybrid","On_Premises")],by = c("COUNTY" = "county")) %>%
  mutate(death_prop = CUMDEATHS/POPULATION)

# plot death prop over time by the majority teaching method
case_policy_wide%>%
  group_by(DATE, major_teaching) %>%
  drop_na(major_teaching)%>%
  summarise(total_deaths = sum(CUMDEATHS),
            total_pop = sum(POPULATION),
            death_prop = total_deaths/total_pop,
            death_prop_upper = death_prop + z_cl*sqrt(death_prop*(1 - death_prop)/total_pop),
            death_prop_lower = death_prop - z_cl*sqrt(death_prop*(1 - death_prop)/total_pop),
            .groups = "drop") %>%
  ggplot(aes(x = DATE, y = death_prop*1000, group = major_teaching))+
  geom_rect(data=case_policy_wide[1,],
            aes(xmin=as.Date("2020/08/18"), xmax=as.Date("2020/12/15"),
                ymin=-Inf,ymax=Inf),
            color = NA,alpha=0.2, show.legend = F, fill = "orange") + 
  geom_line(aes(color = major_teaching),size = 1, alpha = .8) + 
  geom_ribbon(aes(ymin = 1000*death_prop_lower, ymax = 1000*death_prop_upper,
                  fill= major_teaching), 
              alpha = .3, show.legend = F)+ 
  geom_vline(xintercept = date.intercept, linetype = "dashed") + 
  annotate("text",x = date.intercept,y = 1.5,
           label = date.intercept,
           hjust = 1.1) + 
  theme_bw() + 
  ggtitle("Death Incidences Increase Faster for Red Counties ")+
  labs(x = "Date", y = "Cumulative Death Incidence / 1,000 people",
       subtitle = "Yellow area represents Fall Semester",color = "Majority Teaching Method") +
  theme(legend.position = "")+theme(legend.title = element_text(size=13),legend.text = element_text(size=13),axis.title = element_text(size=14),axis.text = element_text(size=15),legend.background = element_rect(fill = alpha("orange",0.0)),legend.key.size = unit(1.4,"lines"),title = element_text(size=12.9))+
  theme(axis.line = element_line(colour = "black"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_blank(),
    panel.background = element_blank()) 

ggsave("totaldeath.jpg", width = 6,height = 5)

```


```{r}
######################### Boxplots ############

# join most common open date for each county
opendate_cases <- case_policy_wide%>%
  inner_join(major_reopening%>%select(COUNTY,major_opendate),by=c('COUNTY'))


# Box Plots in Fall semester
library(PMCMRplus)
require(DescTools)

fall_cases <- opendate_cases %>%
  filter(DATE >= major_opendate & DATE <= as.Date("2020/12/15")) %>%
  group_by(COUNTY) %>%
  arrange(DATE) %>%
  filter(row_number()==1 | row_number()==n()) %>%
  mutate(death_incidence = diff(CUMDEATHS),
         death_incidence_per_1000 = death_incidence*1000/POPULATION) %>%
  distinct(COUNTY,POPULATION,major_teaching,
           death_incidence,death_incidence_per_1000) 

fall_major_teaching.aov <- aov(death_incidence_per_1000 ~ major_teaching,data = fall_cases)
summary(fall_major_teaching.aov) # p-value of .012
stat.test <- PostHocTest(fall_major_teaching.aov, method = "duncan")$major_teaching %>%
  as.data.frame()%>%
  rownames_to_column("group") %>%
  separate(group,"-", into = c("group1","group2")) %>%
  mutate(pval = round(pval,3),
         p = case_when(pval <= .01~ "**",
                       pval <= .05 ~ "*",
                       TRUE ~ "NS"))%>%
  select(group1, group2, pval, p)

library(ggpubr)

ggplot(fall_cases,aes(y = death_incidence_per_1000, x = major_teaching)) + 
  geom_boxplot(aes(fill = major_teaching))+
  stat_compare_means(method = "anova")+ 
  stat_pvalue_manual(stat.test, label = "p",y.position = 2.5, step.increase = 0.15)+ 
  ylim(c(0,3.5))+ 
  theme_bw()+ 
  labs(y = "Death Incidence / 1,000 people",
       fill = "Majority Teaching Method",
       title = "Death Incidence in the Fall Semester",
       caption = "Pairwise p-values come from Duncan pairwise comparison test") +
  theme(legend.position = "bottom",
        axis.text.x=element_blank())

#ggsave("fall_boxplots.jpg",width = 8.5, height = 5)


# Assign 3 key windows
interval_cases <- opendate_cases %>%
  mutate(interval = case_when(DATE >= as.Date(major_opendate) &
                     DATE <= as.Date(major_opendate) + 21 ~ "Start of School - 3 weeks after Start of School",
                   DATE >= as.Date("2020/11/24") - 10 & 
                     DATE<= as.Date("2020/11/24") + 11 ~ "11/14/2020 - 12/05/2020",
                   DATE >= as.Date("2020/12/15") + 21 &
                     DATE <= as.Date("2020/12/15") + 42 ~ "01/05/2021 - 01/26/2020"))%>%
  group_by(COUNTY, interval) %>%
  arrange(DATE) %>%
  filter(row_number()==1 | row_number()==n())%>%
  mutate(death_incidence = diff(CUMDEATHS),
         death_incidence_per_1000 = death_incidence*1000/POPULATION) %>%
  distinct(COUNTY,POPULATION,major_teaching,
           death_incidence,death_incidence_per_1000,interval) %>%
  filter(!is.na(interval)) 

interval_cases$interval <- factor(interval_cases$interval, 
                                  levels = c("Start of School - 3 weeks after Start of School",
                                             "11/14/2020 - 12/05/2020",
                                             "01/05/2021 - 01/26/2020"))

# Box Plots in 3 key Windows 
ggplot(interval_cases, aes(y = death_incidence_per_1000, 
                           x = major_teaching,fill = major_teaching)) + 
  geom_boxplot() +
  stat_compare_means(method = "anova")+ 
  facet_wrap(~interval) + 
  theme_bw() + 
  labs(y = "Death Incidence / 1,000 people",
       fill = "Majority Teaching Method",
       title = "Death Incidence in 3 Key Windows",
       subtitle = "3 Week Windows",
       caption = "The first window looks at the 3 weeks after the start of school, \nthe second is the widndow around the point of intersection, \nand the last window is 3 weeks after the end of the fall semester.") + 
  theme(legend.position = "bottom",
        axis.text.x=element_blank())


#ggsave("int_boxplots.jpg",width = 8.5, height = 5)

```



```{r}
# Derivative Plot 

lag_cases <- case_mobility %>%
  left_join(county_policy_wide[,c("county","major_teaching")],
            by = c("COUNTY" = "county")) %>%
  drop_na(major_teaching)%>%
  select(COUNTY,DATE,CUMDEATHS,POPULATION,major_teaching)%>%
  group_by(COUNTY) %>%
  mutate(lag_total_deaths = lag(CUMDEATHS,21)) %>%
  ungroup()%>%
  group_by(DATE,major_teaching) %>%
  summarise(total_deaths = sum(CUMDEATHS),
            total_deaths_lag = sum(lag_total_deaths),
            total_pop = sum(POPULATION),
            death_prop = total_deaths/total_pop,
            lag_death_prop = total_deaths_lag/total_pop,
            death_prop_inc = (total_deaths-total_deaths_lag)/total_pop,
            .groups = "drop")

peak.date <- as.Date("2020-12-23")

ggplot(lag_cases,aes(x = DATE, y = death_prop_inc*1000, 
                     group = major_teaching)) + 
  geom_line(na.rm = T, aes(color = major_teaching)) +
  geom_rect(data = lag_cases[1,],
            aes(xmin=as.Date("2020/08/26"), xmax=as.Date("2020/12/12"),
                ymin=-Inf,ymax=Inf),
            color = NA,alpha=0.2, show.legend = F, fill = "orange") + 
  geom_vline(xintercept = peak.date, linetype = "dashed")+ 
  annotate("text",x = peak.date,y = .0005,
           label = peak.date,
           hjust = 1.2)  + 
  theme_bw() + 
  labs(x = "Date",
       y = "Death Proportion / 1,000 people", 
       title = "Death Proportion Increase by Teaching Method",
       subtitle = "Yellow area represents Fall Semester",
       caption = "Increase compared to 3 Week Lag",
       color = "Majority Teaching Method") + 
  scale_y_continuous(labels = comma)

```


```{r}
library(ggrepel)
library(cowplot)
library(sp)
ohio_map <- map_data("county") %>%subset(region=="ohio")%>%
  mutate(county=toupper(subregion))%>%select(long,lat,county,group)

# create map plots
wide_teaching_enroll%>%
  left_join(ohio_map,by='county')%>%
  mutate(Online_Only= Online_Only*100)%>%
  ggplot() + geom_polygon(aes(x = long, y = lat, group = group, fill = Online_Only), color = "gray") + coord_fixed(1.3) + theme_map() +
scale_fill_distiller(palette = "OrRd",direction = 1)+labs(fill='% Online Only')+theme(legend.text = element_text(size=20),legend.title = element_text(size=20))



# create map plots
wide_teaching_enroll%>%
  left_join(ohio_map,by='county')%>%
  mutate(On_Premises= On_Premises*100)%>%
  ggplot() + geom_polygon(aes(x = long, y = lat, group = group, fill = On_Premises), color = "gray") + coord_fixed(1.3) + theme_map() +
scale_fill_distiller(palette = "OrRd",direction = 1)+labs(fill='% On Premises')+theme(legend.text = element_text(size=20),legend.title = element_text(size=20))


# create map plots for population
wide_teaching_enroll%>%
  left_join(ohio_map,by='county')%>%
  mutate(Hybrid= Hybrid*100)%>%
  ggplot() + geom_polygon(aes(x = long, y = lat, group = group, fill = Hybrid), color = "gray") + coord_fixed(1.3) + theme_map() +
scale_fill_distiller(palette = "OrRd",direction = 1)+labs(fill='% Hybrid')+theme(legend.text = element_text(size=20),legend.title = element_text(size=20))



# create map plots
cases%>%
  distinct(COUNTY,POPULATION)%>%
  left_join(ohio_map,by=c('COUNTY'='county'))%>%
  mutate(population = POPULATION/1000)%>%
  ggplot() + geom_polygon(aes(x = long, y = lat, group = group, fill = population), color = "gray") + coord_fixed(1.3) + theme_map() +
scale_fill_distiller(palette = "OrRd",direction = 1)+labs(fill='Population/1000')+theme(legend.text = element_text(size=12),legend.title = element_text(size=12),legend.position = "bottom",legend.key.size = unit(2,"lines"))

# create map plots
teachingmethod_enroll%>%
  distinct(county,county_enroll)%>%
  left_join(ohio_map,by=c('county'))%>%
  mutate(county_enroll = county_enroll/1000)%>%
  ggplot() + geom_polygon(aes(x = long, y = lat, group = group, fill = county_enroll), color = "gray") + coord_fixed(1.3) + theme_map() +
scale_fill_distiller(palette = "OrRd",direction = 1)+labs(fill='Enrollment/1000')+theme(legend.text = element_text(size=12),legend.title = element_text(size=12),legend.position = "bottom",legend.key.size = unit(2,"lines"))

```


```{r}
date.intercept

# aggregate daily mobility at state level by teaching method
date_mobility

# reorder the level 
date_mobility$major_teaching <- factor(date_mobility$major_teaching,
                                            levels = c("On Premises",
                                                       "Hybrid",
                                                       "Online Only"))
# FULL TIME WORK, 6 hours +
date_mobility%>%
  ggplot(aes(x=DATE,y=full_work_prop))+
  geom_rect(data=date_mobility[1,],
            aes(xmin=as.Date("2020/08/26"), xmax=as.Date("2020/12/12"),
                ymin=-Inf,ymax=Inf),
            color = NA,alpha=0.2, show.legend = F, fill = "orange") + 
  geom_line(aes(color = major_teaching),size = 1, alpha = .8)+
  geom_vline(xintercept = date.intercept, linetype = "dashed") + 
  annotate("text",x = date.intercept,y = 0.08,
           label = date.intercept,
           hjust = 1.1,size=5) + 
  theme_bw() + 
  labs(x = "Date", y = "% Cell phones in population",caption = "Average over 7 days",subtitle = "Yellow area represents Fall Semester",
       color = "Majority Teaching Method",title = "Percent Cell Phones Away Home for 6hr+") + ylim(0.02,0.10)+
  theme(legend.position = "right")+theme(legend.title = element_text(size=13),legend.text = element_text(size=13),axis.title = element_text(size=15),axis.text = element_text(size=15),legend.background = element_rect(fill = alpha("orange",0.0)),legend.key.size = unit(1.4,"lines"),title = element_text(size=12.9))+
  theme(axis.line = element_line(colour = "black"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_blank(),
    panel.background = element_blank()) 
```


```{r}
date_mobility%>%
  ggplot(aes(x=DATE,y=part_work_prop))+
  geom_rect(data=date_mobility[1,],
            aes(xmin=as.Date("2020/08/26"), xmax=as.Date("2020/12/12"),
                ymin=-Inf,ymax=Inf),
            color = NA,alpha=0.2, show.legend = F, fill = "orange") + 
  geom_line(aes(color = major_teaching),size = 1, alpha = .8)+
  geom_vline(xintercept = date.intercept, linetype = "dashed") + 
  annotate("text",x = date.intercept,y = 0.1,
           label = date.intercept,
           hjust = 1.1) + 
  theme_bw() + 
  labs(x = "Date", y = "% Devices in population",subtitle = "Average over 7 days",
       color = "Majority Teaching Method",title = "Percentage of Devices Away Home for 3-6hr") +
  theme(legend.position = "bottom")+theme(legend.title = element_text(size=13),legend.text = element_text(size=13),axis.title = element_text(size=15),axis.text = element_text(size=15),legend.background = element_rect(fill = alpha("orange",0.0)),legend.key.size = unit(1.4,"lines"),title = element_text(size=12.9))
```

```{r}
# plot death prop over time by the majority teaching method
case_policy_wide%>%
  group_by(DATE, major_teaching) %>%
  drop_na(major_teaching)%>%
  summarise(total_deaths = sum(CUMDEATHS),
            total_pop = sum(POPULATION),
            death_prop = total_deaths/total_pop,
            death_prop_upper = death_prop + z_cl*sqrt(death_prop*(1 - death_prop)/total_pop),
            death_prop_lower = death_prop - z_cl*sqrt(death_prop*(1 - death_prop)/total_pop),
            .groups = "drop") %>%
  ggplot(aes(x = DATE, y = death_prop*1000, group = major_teaching))+
  geom_rect(data=case_policy_wide[1,],
            aes(xmin=as.Date("2020/08/26"), xmax=as.Date("2020/12/12"),
                ymin=-Inf,ymax=Inf),
            color = NA,alpha=0.2, show.legend = F, fill = "orange") + 
  geom_line(aes(color = major_teaching),size = 1, alpha = .8) + 
  geom_ribbon(aes(ymin = 1000*death_prop_lower, ymax = 1000*death_prop_upper,
                  fill= major_teaching), 
              alpha = .3, show.legend = F)+ 
  geom_vline(xintercept = date.intercept, linetype = "dashed") + 
  annotate("text",x = date.intercept,y = 1.5,
           label = date.intercept,
           hjust = 1.1) + 
  theme_bw() + 
  labs(x = "Date", y = "Cumulative Death Incidence / 1,000 people",
       color = "Majority Teaching Method") + 
  theme(legend.position = "bottom")+theme(legend.title = element_text(size=13),legend.text = element_text(size=13),axis.title = element_text(size=15),axis.text = element_text(size=15),legend.background = element_rect(fill = alpha("orange",0.0)),legend.key.size = unit(1.4,"lines"),title = element_text(size=12.9))+geom_line(data=date_mobility,aes(x=DATE,y=full_work_prop,color=major_teaching))

```


```{r}
date_mobility%>%
  ggplot(aes(x=DATE,y=res_visit_prop))+
  geom_rect(data=date_mobility[1,],
            aes(xmin=as.Date("2020/08/26"), xmax=as.Date("2020/12/12"),
                ymin=-Inf,ymax=Inf),
            color = NA,alpha=0.2, show.legend = F, fill = "orange") + 
  geom_line(aes(color = major_teaching),size = 1, alpha = .8)+
  geom_vline(xintercept = date.intercept, linetype = "dashed") + 
  annotate("text",x = date.intercept,y = 0.1,
           label = date.intercept,
           hjust = 1.1) + 
  theme_bw() + 
  labs(x = "Date", y = "Number of Restaurant Visit",subtitle = "Average over 7 days, no data for Hybrid counties",
       color = "Majority Teaching Method",title = "Number of Restaurant Visit Normalized by Population") +
  theme(legend.position = "bottom")+theme(legend.title = element_text(size=13),legend.text = element_text(size=13),axis.title = element_text(size=15),axis.text = element_text(size=15),legend.background = element_rect(fill = alpha("orange",0.0)),legend.key.size = unit(1.4,"lines"),title = element_text(size=12.9))
```

```{r}
state_cases <- cases%>%
  group_by(DATE)%>%
  summarise(state_death = sum(CUMDEATHS,na.rm = T),state_new_death=sum(NEWDEATHS,na.rm = T))

state_cases%>%
  ggplot(aes(x = DATE, y = state_death))+geom_line(size = 1)+
  geom_rect(data=state_cases[1,],
            aes(xmin=as.Date("2020/08/18"), xmax=as.Date("2020/12/12"),
                ymin=-Inf,ymax=Inf),
            color = NA,alpha=0.2, show.legend = F, fill = "orange") + 
  geom_vline(xintercept = as.Date("2020/11/24"), linetype = "dashed") + 
  annotate("text",x = as.Date("2020/11/24"),y = 12000,
           label = "2020/11/24",
           hjust = 1.1) + 
  theme_bw() + 
  labs(x = "Date", y = "Cumulative Death Incidence") + 
  theme(legend.position = "bottom")+theme(,axis.title = element_text(size=15),axis.text = element_text(size=15),title = element_text(size=12.9))


state_cases%>%
  filter(DATE<as.Date("2021/01/31"))%>%
  ggplot(aes(x = DATE, y = state_new_death))+geom_line(size = 1)+
  geom_rect(data=state_cases[1,],
            aes(xmin=as.Date("2020/08/18"), xmax=as.Date("2020/12/12"),
                ymin=-Inf,ymax=Inf),
            color = NA,alpha=0.2, show.legend = F, fill = "orange") + 
  geom_vline(xintercept = as.Date("2020/11/24"), linetype = "dashed") + 
  annotate("text",x = as.Date("2020/11/24"),y = 200,
           label = "2020/11/24",
           hjust = 1.1) + 
  theme_bw() + ylim(0,300)+
  labs(x = "Date", y = "New Death Incidence") + 
  theme(legend.position = "bottom")+theme(,axis.title = element_text(size=15),axis.text = element_text(size=15),title = element_text(size=12.9))
```




```{r}
# create map plots
ohio_profile%>%
  left_join(ohio_map,by=c('County'='county'))%>%
  ggplot() + geom_polygon(aes(x = long, y = lat, group = group, fill = as.factor(NCHS.Urban.Rural.Status)), color = "gray") + coord_fixed(1.3) + theme_map() +labs(fill='Urban Rural Status')+theme(legend.text = element_text(size=20),legend.title = element_text(size=20))+
  scale_fill_brewer(palette = "Reds",direction = -1)+theme(legend.title = element_text(size=16),legend.text = element_text(size=16),legend.key.size = unit(1.4,"lines"))

```


```{r}
names(ohio_profile)
# create map plots
ohio_profile%>%
  left_join(ohio_map,by=c('County'='county'))%>%
  ggplot() + geom_polygon(aes(x = long, y = lat, group = group, fill = as.factor(Metropolitan.Status)), color = "gray") + coord_fixed(1.3) + theme_map() +labs(fill='Metropolitan Status')+theme(legend.text = element_text(size=20),legend.title = element_text(size=20))+scale_fill_manual(values = c("tomato3","seashell"))+theme(legend.title = element_text(size=13),legend.text = element_text(size=13),legend.key.size = unit(1.4,"lines"))

```
