==> RawCodes/step11_new deaths plot.Rmd <==
---
title: "new deaths plot"
author: "Yixuan Luo"
date: "4/15/2021"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## New Deaths Plot

```{r}
county_policy_wide$major_teaching <- factor(county_policy_wide$major_teaching,levels = c("On Premises",
                                                       "Hybrid",
                                                      "Online Only"))
# see when the intesection happens
date.intercept <- as.Date("2020-11-20")
# add 95% confidence bans
confidence_level <- .95
z_cl <- qnorm(confidence_level)
# case_policy_wide
case_policy_wide <- cases %>%
  left_join(county_policy_wide[,c("county","major_teaching","Online_Only","Hybrid","On_Premises")],by = c("COUNTY" = "county")) %>%
  mutate(death_prop = CUMDEATHS/POPULATION)
```

```{r}
case_policy_wide <- case_policy_wide %>%
  mutate(oo_death = Online_Only*CUMDEATHS) %>%
  mutate(hy_death = Hybrid*CUMDEATHS) %>%
  mutate(op_death = On_Premises*CUMDEATHS)
```

```{r}
case_policy_wide <- case_policy_wide %>%
  mutate(oo_adjust = Online_Only*POPULATION) %>%
  mutate(hy_adjust = Hybrid*POPULATION) %>%
  mutate(op_adjust = On_Premises*POPULATION)
```

```{r}
case_policy_sum <- case_policy_wide %>%
  group_by(DATE) %>%
  summarise(oo_death_sum = sum(oo_death, na.rm = T),
            hy_death_sum = sum(hy_death, na.rm = T),
            op_death_sum = sum(op_death, na.rm = T),
            oo_adjust_sum = sum(oo_adjust, na.rm = T),
            hy_adjust_sum = sum(hy_adjust, na.rm = T),
            op_adjust_sum = sum(op_adjust, na.rm = T))
```

```{r}
case_policy_sum <- case_policy_sum %>%
  mutate(Online_Only = oo_death_sum/oo_adjust_sum) %>%
  mutate(Hybrid = hy_death_sum/hy_adjust_sum) %>%
  mutate(On_Premises = op_death_sum/oo_adjust_sum)
```

```{r}
case_policy_try <- case_policy_sum %>%
  select(DATE, On_Premises, Hybrid, Online_Only)
```

```{r}
case_policy_try <- melt(case_policy_try, "DATE")
```

```{r}
# plot death prop over time by the majority teaching method
case_policy_try %>%
  ggplot(aes(x=DATE,y=value,color=variable,group=variable))+
  geom_rect(data=case_policy_try[1,],
            aes(xmin=as.Date("2020/08/18"), xmax=as.Date("2020/12/15"),
                ymin=-Inf,ymax=Inf),
            color = NA,alpha=0.2, show.legend = F, fill = "orange") + 
  geom_line(size = 1, alpha = .8)+
  geom_vline(xintercept = date.intercept, linetype = "dashed") + 
  annotate("text",x = date.intercept,y = 0.0015,
           label = date.intercept,
           hjust = 1.1) + 
  theme_bw() + 
  ggtitle("Death Incidences by Teaching Method (aggregated by proportions of teaching methods)")+
  labs(x = "Date", y = "Cumulative Death Incidence",
       subtitle = "Yellow area represents Fall Semester",color = "Teaching Method") +
  theme(legend.position = "")+theme(legend.title = element_text(size=13),legend.text = element_text(size=13),axis.title = element_text(size=14),axis.text = element_text(size=15),legend.background = element_rect(fill = alpha("orange",0.0)),legend.key.size = unit(1.4,"lines"),title = element_text(size=12.9))+
  theme(axis.line = element_line(colour = "black"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_blank(),
    panel.background = element_blank()) 
ggsave("totaldeath_new.jpg",width = 10,height = 7)

```


#simple version to check labels
```{r}
ggplot(case_policy_try,aes(x=DATE,y=value,colour=variable,group=variable)) + geom_line()+theme_bw()
```
