==> RawCodes/step12_B_logdeaths_April14.rmd <==
---
title: "B vs. Log Deaths"
author: "Cheyenne, Ziyan"
date: "4/11/2021"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Notes:

The 1st derivative of log new deaths+1 we observed on date T corresponds to the death growth rate B happend on date T-24.

The mobility date has been shifted one week forward as well. Assuming mobility a week before date T change the infections on date T

# Load in the spline slopes

```{r message=FALSE, warning=FALSE, cache=TRUE}
source("../Data_processor/007_B_slopes_processor.R")

# Aggregate counties by teaching method

cases_slope_teach_agg <- cases_slope_teach %>%
  drop_na(major_teaching)%>%
  group_by(DATE, major_teaching) %>%
  summarise(total_new_deaths = sum(rev_NEWDEATHS), .groups = "drop") %>%
  mutate(log_new_deaths = log(total_new_deaths + 1)) %>%
  group_by(major_teaching) %>%
  mutate(smooth.spline = smooth.spline(DATE,log_new_deaths,df = 398/28)$y,
         B = predict(smooth.spline(DATE,log_new_deaths,df = 398/28),deriv = 1)$y)

```




```{r}
# maxB1 vs avg1w_2w_B0
B0B1%>%
  drop_na(major_teaching)%>%
  ggplot(aes(x=avg1w_2w_B0,y=max_B1,group=major_teaching,color=major_teaching))+geom_point(aes(size=avg_full_work_prob,alpha=major_teaching_prop))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+labs(color="Majority Teaching Method",size="Averaged \n%6hr+ Away Home",title="Max B1 in Fall v.s.\n averaged B0 between 1 week before and 2 weeks after reopen",alpha="Size of Majority",y="Max B1 in Fall", x= "Averaged B0 in 3 weeks")+
  team_theme+team_theme+geom_vline(xintercept = 0, lty = 2)
#ggsave("maxB1vsAvgB01w2w.png",width = 7, height = 7)

# avgB1 vs avg1w_2w_B0
B0B1%>%
  drop_na(major_teaching)%>%
  ggplot(aes(x=avg1w_2w_B0,y=avg_B1,group=major_teaching,color=major_teaching))+geom_point(aes(size=avg_full_work_prob,alpha=major_teaching_prop))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+labs(color="Majority Teaching Method",size="Averaged \n%6hr+ Away Home",title="Averaged B1 in Fall v.s.\n averaged B0 1 week before and 2 weeks after reopen",alpha="Size of Majority",y="Averaged B1 in Fall", x= "Averaged B0 in 3 weeks")+
  team_theme+team_theme+geom_vline(xintercept = 0, lty = 2)
#ggsave("avgB1vsAvgB01w2w.png",width = 7, height = 7)
```


```{r  message=FALSE, warning=FALSE}
# max B1 vs avg3w_B0
B0B1%>%
  ggplot(aes(x=avg3w_B0,y=max_B1,group=major_teaching,color=major_teaching))+geom_point(aes(size=avg_full_work_prob,alpha=major_teaching_prop))+theme_minimal()+labs(color="Majority Teaching Method",size="Averaged \n%6hr+ Away Home",title="max B1 v.s. averaged B0 in 3 weeks after reopen",alpha="Size of Majority",y="Max B1", x= "Averaged B0 in 3 weeks")+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+geom_vline(xintercept = 0, lty = 2)+team_theme
ggsave("maxB1vsAvgB0.png",width = 7, height = 7)

# avgB1 vs avg3w_B0
B0B1%>%
  ggplot(aes(x=avg3w_B0,y=avg_B1,group=major_teaching,color=major_teaching))+geom_point(aes(size=avg_full_work_prob,alpha=major_teaching_prop))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+labs(color="Majority Teaching Method",size="Averaged \n%6hr+ Away Home",title="Averaged B1 in Fall v.s. averaged B0 in 3 weeks after reopen",alpha="Size of Majority",y="Averaged B1 in Fall", x= "Averaged B0 in 3 weeks")+
  team_theme+team_theme+geom_vline(xintercept = 0, lty = 2)
ggsave("avgB1vsAvgB0.png",width = 7, height = 7)
```


## Max B1 vs Death
```{r}
death_semester1<-cases%>%
  filter(DATE=="2020-08-18")%>%
  select(COUNTY,CUMDEATHS)
names(death_semester1)[2]<-"start"

death_semester2<-cases%>%
  filter(DATE=="2020-12-15")%>%
  select(COUNTY,CUMDEATHS)
names(death_semester2)[2]<-"end"

death_semester1<-death_semester1%>%left_join(death_semester2,by="COUNTY")%>%mutate(death_semester=end-start)

# Max B1 vs Log of Deaths during fall semester
na.omit(B0B1)%>%
  left_join(death_semester1,by="COUNTY")%>%
  ggplot(aes(x=log(death_semester),y=max_B1,group=major_teaching,color=major_teaching))+geom_point(aes(size=avg_full_work_prob))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+labs(y="Max B1",x="Log of Deaths during fall semester",title="Max B1 vs Log of Deaths during fall semester",size="Averaged \n%6hr+ Away Home")
ggsave("maxB1vslogdeathsem.png",width = 7, height = 5)

# Max B1 vs Log of cumulative Deaths - the cut date(end of semester)
na.omit(B0B1)%>%
  left_join(death_semester1,by="COUNTY")%>%
  ggplot(aes(x=log(end),y=max_B1,group=major_teaching,color=major_teaching))+geom_point(aes(size=avg_full_work_prob))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+labs(y="Max B1",x="Log of total deaths",title="Max B1 vs Log of total deaths (until 2020-12-15)",size="Averaged \n%6hr+ Away Home")
ggsave("maxB1vslogdeath.png",width = 7, height = 5)

# Max B1 vs Log of cumulative Deaths - the cut date(2021-2-22)
na.omit(B0B1)%>%
  left_join(death_prop%>%select(COUNTY,CUMDEATHS),by="COUNTY")%>%
  ggplot(aes(x=log(CUMDEATHS),y=max_B1,group=major_teaching,color=major_teaching))+geom_point(aes(size=avg_full_work_prob))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+labs(y="Max B1",x="Log of total deaths",title="Max B1 vs Log of total deaths (until 2021-02-22)",size="Averaged \n%6hr+ Away Home")
ggsave("maxB1vslogdeath_222.png",width = 7, height = 5)
```



```{r}
# Max B1 vs Log of Deaths rate during fall semester
na.omit(B0B1)%>%
  left_join(death_semester1,by="COUNTY")%>%
  left_join(county_population,by="COUNTY")%>%
  mutate(log_death_rate=log(death_semester/POPULATION))%>%
  ggplot(aes(x=log_death_rate,y=max_B1,group=major_teaching,color=major_teaching))+geom_point(aes(size=avg_full_work_prob))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+labs(y="Max B1",x="Log of Death Rate",title="Max B1 vs Log of Death Rate during fall semester-log(death/population)",size="Averaged \n%6hr+ Away Home")
ggsave("maxB1vslogdeathrate1.png",width = 7, height = 5)

na.omit(B0B1)%>%
  left_join(death_semester1,by="COUNTY")%>%
  left_join(county_population,by="COUNTY")%>%
  mutate(log_death_rate=log(death_semester/POPULATION+1))%>%
  ggplot(aes(x=log_death_rate,y=max_B1,group=major_teaching,color=major_teaching))+geom_point(aes(size=avg_full_work_prob))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+labs(y="Max B1",x="Log of Death rate",title="Max B1 vs Log of Death Rate during fall semester-log(death/population+1)",size="Averaged \n%6hr+ Away Home")
ggsave("maxB1vslogdeathrate2.png",width = 7, height = 5)

na.omit(B0B1)%>%
  left_join(death_semester1,by="COUNTY")%>%
  left_join(county_population,by="COUNTY")%>%
  mutate(log_death_rate=death_semester/POPULATION*1000)%>%
  ggplot(aes(x=log_death_rate,y=max_B1,group=major_teaching,color=major_teaching))+geom_point(aes(size=avg_full_work_prob))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+labs(y="Max B1",x="Deaths per 1000 people",title="Max B1 vs Death Rate during fall semester-Deaths per 1000 people",size="Averaged \n%6hr+ Away Home")
ggsave("maxB1vslogdeathrate3.png",width = 7, height = 5)
```


```{r}
# Max B1 vs Log of total death rate 
na.omit(B0B1)%>%
  left_join(death_semester1,by="COUNTY")%>%
  left_join(county_population,by="COUNTY")%>%
  mutate(log_death_rate=log(end/POPULATION))%>%
  ggplot(aes(x=log_death_rate,y=max_B1,group=major_teaching,color=major_teaching))+geom_point(aes(size=avg_full_work_prob))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+labs(y="Max B1",x="Log of Total Death Rate",title="Max B1 vs Log of Total Death Rate-log(death/population)",size="Averaged \n%6hr+ Away Home")
ggsave("maxB1vstotaldeathrate.png",width = 7, height = 5)
```




```{r}
mobility_semester2<-case_mobility%>%
  filter(DATE<=as.Date("2020-12-15")&DATE>=as.Date("2020-08-18"))%>%
  select(COUNTY,full_work_prop_7d,POPULATION)%>%
  group_by(COUNTY)%>%
  summarise(total_mobility=sum(full_work_prop_7d*POPULATION),
            POPULATION=POPULATION)%>%
  distinct(COUNTY,total_mobility,POPULATION)


na.omit(B0B1)%>%
  left_join(mobility_semester2,by="COUNTY")%>%
  ggplot(aes(x=log(total_mobility),y=max_B1,group=major_teaching,color=major_teaching))+geom_point(aes(size=log(POPULATION)))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+labs(y="Max B1",x="Log of Total mobility during fall semester",title="Max B1 vs Log of Total Mobility during fall semester \n- sum of full work proportion times population")

ggsave("maxB1vstotalmob.png",width = 7, height = 5)
```



## Cumulative mobility

Plot1.1: cumulative mobility per color (plot the 3 curves on the same plot).

Aggregate counties by teaching method

```{r}
cum_mobility <- case_mobility%>%
  left_join(wide_teaching_enroll,by=c("COUNTY"="county"))%>%
  left_join(major_reopening,by="COUNTY")%>%
  group_by(DATE,major_teaching)%>%
  summarise(daily_tot_teaching_full_work= sum(full_work_prop_7d*POPULATION),
            tot_teaching_pop = sum(POPULATION), .groups = "drop")%>%
  group_by(major_teaching)%>%
  arrange(DATE)%>%
  mutate(cum_full_work = cumsum(daily_tot_teaching_full_work),cum_full_work_prop = cumsum(daily_tot_teaching_full_work)/tot_teaching_pop)%>%
  mutate(shiftedDate = DATE+28) ## Shifted Dates!!!!!!!!!!!!
cum_mobility$major_teaching <- factor(cum_mobility$major_teaching,levels = c("On Premises","Hybrid","Online Only"))
```

```{r}
cum_mobility_semester<-wide_teaching_enroll%>%
  left_join(case_mobility,by=c("county"="COUNTY"))%>%
  left_join(major_reopening,by=c("county"="COUNTY"))%>%
  filter(DATE>=as.Date("2020-08-18") &DATE<=as.Date("2020-12-15"))%>%
  group_by(county,DATE)%>%
  summarise(cum_full_work = sum(full_work_prop_7d*POPULATION),cum_full_work_prop = sum(full_work_prop_7d*POPULATION)/sum(POPULATION), .groups = "drop")%>%
  mutate(shiftedDate = DATE+28)

cum_mobility_semester1<-cum_mobility_semester%>%
  filter(DATE)
na.omit(B0B1)%>%
  left_join(mobility_semester,by="COUNTY")%>%
  ggplot(aes(x=total_mobility,y=max_B1,group=major_teaching,color=major_teaching))+geom_point()+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+labs(y="Max B1",x="Average mobility during fall semester",title="Max B1 vs Average full mobility during fall semester")
```

```{r}
options(scipen=1000)
ggplot(cum_mobility, aes(x = shiftedDate, color = major_teaching)) +
  geom_line(aes(y = cum_full_work), alpha = 1,size=1)+ 
  geom_rect(data = cum_mobility[1,],
            aes(xmin=as.Date("2020/08/18"), xmax=as.Date("2020/12/15"),
                ymin=-Inf,ymax=Inf),
            color = NA,alpha=0.2, show.legend = F, fill = "orange") + 
  theme_bw() + 
  labs(x = "Date", y = "Cumulative Full-time Workers/Students", 
       color = "Majority Teaching Method", title="Cumulative Full-time Mobility (6hrs+ Away Home)\n4 Weeks Forward",
       subtitle = "Yellow area represents the fall semester (08/18 - 12/15)") + 
  theme(axis.line = element_line(colour = "black"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_blank(),
    panel.background = element_blank(),legend.position = "bottom")+theme(legend.text = element_text(size=12),legend.title = element_text(size=12),axis.text = element_text(size=13),title=element_text(size=13))
#ggsave("cum_mobility_time_series.png",width = 7, height = 5)
```

Plot1.2: average mobility per color (plot the 3 curves on the same plot).

```{r}
date.intercept=as.Date("2020-11-24")
date_mobility$major_teaching <- factor(date_mobility$major_teaching,levels = c("On Premises","Hybrid","Online Only"))
date_mobility$shiftedDate <- date_mobility$DATE + 28
# FULL TIME WORK, 6 hours +
date_mobility%>%
  ggplot(aes(x=shiftedDate,y=full_work_prop))+
  geom_rect(data=date_mobility[1,],
            aes(xmin=as.Date("2020/08/26"), xmax=as.Date("2020/12/12"),
                ymin=-Inf,ymax=Inf),
            color = NA,alpha=0.2, show.legend = F, fill = "orange") + 
  geom_line(aes(color = major_teaching),size = 1, alpha = .8)+
  geom_vline(xintercept = date.intercept, linetype = "dashed") + 
  annotate("text",x = date.intercept,y = 0.08,
           label = date.intercept,
           hjust = 1.1,size=5) + 
  theme_bw() + 
  labs(x = "Date", y = "% Full-time Workers/Students",subtitle = "Average over 7 days; Yellow area represents Fall Semester",
       color = "Majority Teaching Method",title = "Average %Cell Phones Away Home for 6hr+\n4 Weeks Forward") + ylim(0.02,0.10)+
  theme(legend.position = "bottom")+theme(legend.title = element_text(size=13),legend.text = element_text(size=13),axis.title = element_text(size=15),axis.text = element_text(size=15),legend.background = element_rect(fill = alpha("orange",0.0)),legend.key.size = unit(1.4,"lines"),title = element_text(size=12.9))+
  theme(axis.line = element_line(colour = "black"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_blank(),
    panel.background = element_blank()) 
```

\newpage

Then a series of 3 plots, each plot showing the first diff of log death and 

cum mobility for each color. If the 2 curves are parallel, then mobility explains death entirely and school has no effect.

Aggregate counties by teaching method (cases_clean)

Plot2.1: On Premises

```{r}
cases_clean_teach_agg%>%
  filter(major_teaching=="On Premises")%>%
  inner_join(cum_mobility%>%filter(major_teaching=="On Premises"),by=c("DATE"="shiftedDate"))%>%
  ggplot(aes(x = DATE))+
  geom_line(aes(y = B),size = 1,color="red") +
  geom_rect(data = cases_clean_teach_agg[1,],
            aes(xmin=as.Date("2020/08/18"), xmax=as.Date("2020/12/15"),
                ymin=-Inf,ymax=Inf),
            color = NA,alpha=0.2, show.legend = F, fill = "orange") + theme_bw() + 
  labs(x = "Date", y = "Exponential Growth Coefficient", title="On-Premises Counties\n Mobility 4 weeks forward",
       color = "Majority Teaching Method",
       caption = "Smoothing window set to every 4 weeks",
       subtitle = "Yellow area represents the fall semester (08/18 - 12/15)") + 
  theme(legend.position = "bottom")+
  geom_line(aes(x = DATE,y=cum_full_work/1000000),color="orange")+
  scale_y_continuous(,# Add a second axis and specify its features
  sec.axis = sec_axis(trans=~.,name="Cumulative Full-time Workers/Students",labels =c("-1"," ","0","25k","50k","75k")))+annotate("text",label="Exponential Growth Coefficient",x=as.Date("2020-09-21"),y=-0.010)+annotate("text",label="Cumulative Full-time Workers/Students",x=as.Date("2020-06-21"),y=0.040)+
  theme(legend.title = element_text(size=13),legend.text = element_text(size=12),axis.title = element_text(size=12),axis.text = element_text(size=12),title = element_text(size=12.9))+
  theme(axis.line = element_line(colour = "black"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_blank(),
    panel.background = element_blank()) 
ggsave("time_series_on_premises.png",width = 7, height = 5)
```

```{r}
cases_clean_teach_agg%>%
  filter(major_teaching=="On Premises")%>%
  inner_join(cum_mobility%>%filter(major_teaching=="On Premises"),by=c("DATE"="shiftedDate"))%>%
  ggplot(aes(y = B,x=cum_full_work))+ 
  geom_point(na.rm = TRUE,colour = 'salmon') + 
  geom_smooth(method = "lm", formula = y~x, se = FALSE, na.rm = TRUE, show.legend = FALSE,colour = 'black') + 
  theme_minimal() + theme(legend.position = "bottom")+
  labs(x = "Cumulative Full-time Workers/Students", y = "Exponential Growth Coefficient", 
       title = "B ~ Cumulative Full-time Workers/Students 4 weeks forward\n On Premises")+theme(text = element_text(size=14),title = element_text(size=13))
ggsave("lm_on_premises.jpg", width = 15, height = 6)
```


Plot2.2: Hybrid
```{r}
cases_clean_teach_agg%>%
  filter(major_teaching=="Hybrid")%>%
  inner_join(cum_mobility%>%filter(major_teaching=="Hybrid"),by=c("DATE"="shiftedDate"))%>%
  ggplot(aes(x = DATE))+
  geom_line(aes(y = B),size = 1,color="darkgreen") +
  geom_rect(data = cases_clean_teach_agg[1,],
            aes(xmin=as.Date("2020/08/18"), xmax=as.Date("2020/12/15"),
                ymin=-Inf,ymax=Inf),
            color = NA,alpha=0.2, show.legend = F, fill = "orange") + theme_bw() + 
  labs(x = "Date", y = "Exponential Growth Coefficient", title="Hybrid Counties\n Mobility 4 weeks forward",
       color = "Majority Teaching Method",
       caption = "Smoothing window set to every 4 weeks",
       subtitle = "Yellow area represents the fall semester (08/18 - 12/15)") + 
  theme(legend.position = "bottom")+
  geom_line(aes(x = DATE,y=cum_full_work/1000000),color="green")+
  scale_y_continuous(# Add a second axis and specify its features
  sec.axis = sec_axis(trans=~.*10,name="Cumulative Full-time Workers/Students",labels = c(" "," ","0","200k","300k","400k","500k")))+
  annotate("text",label="Exponential Growth Coefficient",x=as.Date("2020-09-21"),y=0.06)+
  annotate("text",label="Cumulative Full-time Workers/Students",x=as.Date("2020-06-21"),y=0.40)+
  theme(legend.title = element_text(size=13),legend.text = element_text(size=12),axis.title = element_text(size=12),axis.text = element_text(size=12),title = element_text(size=12.9))+
  theme(axis.line = element_line(colour = "black"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_blank(),
    panel.background = element_blank()) 
#ggsave("time_series_hybrid.png",width = 7, height = 5)
```
```{r}
cases_clean_teach_agg%>%
  filter(major_teaching=="Hybrid")%>%
  inner_join(cum_mobility%>%filter(major_teaching=="Hybrid"),by=c("DATE"="shiftedDate"))%>%
  ggplot(aes(y = B,x=cum_full_work))+ 
  geom_point(na.rm = TRUE,colour = 'darkgreen') + 
  geom_smooth(method = "lm", formula = y~x, se = FALSE, na.rm = TRUE, show.legend = FALSE,colour = 'black') + 
  theme_minimal() + theme(legend.position = "bottom")+
  labs(x = "Cumulative Full-time Workers/Students", y = "Exponential Growth Coefficient", 
       title = "B ~ Cumulative Full-time Workers/Students 4 weeks forward\n Hybrid")+theme(text = element_text(size=14),title = element_text(size=13))
  
```

Plot2.3: Online Only
```{r}
cases_clean_teach_agg%>%
  filter(major_teaching=="Online Only")%>%
  inner_join(cum_mobility%>%filter(major_teaching=="Online Only"),by=c("DATE"="shiftedDate"))%>%
  ggplot(aes(x = DATE))+
  geom_line(aes(y = B),size = 1,color="darkblue") +
  geom_rect(data = cases_clean_teach_agg[1,],
            aes(xmin=as.Date("2020/08/18"), xmax=as.Date("2020/12/15"),
                ymin=-Inf,ymax=Inf),
            color = NA,alpha=0.2, show.legend = F, fill = "orange") + theme_bw() + 
  labs(x = "Date", y = "Exponential Growth Coefficient", title="Online-only Counties\n Mobility 4 weeks forward",
       color = "Majority Teaching Method",
       caption = "Smoothing window set to every 4 weeks",
       subtitle = "Yellow area represents the fall semester (08/18 - 12/15)") + 
  theme(legend.position = "bottom")+
  geom_line(aes(x = DATE,y=cum_full_work/1000000),color="blue")+
  scale_y_continuous(# Add a second axis and specify its features
  sec.axis = sec_axis(trans=~.*10,name="Cumulative Full-time Workers/Students",labels = c(" ","0","100k","200k","300k","400k")))+
  annotate("text",label="Exponential Growth Coefficient",x=as.Date("2020-09-21"),y=0.06)+
  annotate("text",label="Cumulative Full-time Workers/Students",x=as.Date("2020-06-21"),y=0.40)+
  theme(legend.title = element_text(size=13),legend.text = element_text(size=12),axis.title = element_text(size=12),axis.text = element_text(size=12),title = element_text(size=12.9))+
  theme(axis.line = element_line(colour = "black"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_blank(),
    panel.background = element_blank()) 
ggsave("time_series_online_only.png",width = 7, height = 5)
```


```{r}
cases_clean_teach_agg%>%
  filter(major_teaching=="Online Only")%>%
  inner_join(cum_mobility%>%filter(major_teaching=="Online Only"),by=c("DATE"="shiftedDate"))%>%
  ggplot(aes(y = B,x=cum_full_work))+ 
  geom_point(na.rm = TRUE,colour = 'darkblue') + 
  geom_smooth(method = "lm", formula = y~x, se = FALSE, na.rm = TRUE, show.legend = FALSE,colour = 'black') + 
  theme_minimal() + theme(legend.position = "bottom")+
  labs(x = "Cumulative Full-time Workers/Students", y = "Exponential Growth Coefficient", 
       title = "B ~ Cumulative Full-time Workers/Students 4 weeks forward\n Online Only")+theme(text = element_text(size=14),title = element_text(size=13))
  
```

# Max B1 vs Percent children on premises/online/hybrid

```{r}
na.omit(B0B1)%>%
  ggplot(aes(x=Online_Only,y=max_B1,fill="blue"))+geom_point(color="#619CFF")+geom_smooth(method = "lm", se=F, formula = y ~ x, alpha=0.1,color="#619CFF") + theme_minimal()+team_theme+labs(y="Max B1",x="Online Only Proportion",title="Max B1 vs Online Only Proportion")+guides(fill=FALSE)
ggsave("maxB1vsonline.png",width = 7, height = 5)

na.omit(B0B1)%>%
  ggplot(aes(x=Hybrid,y=max_B1))+geom_point(color="#00BA38")+geom_smooth(method = "lm", se=F, formula = y ~ x, alpha=0.1,color="#00BA38") + theme_minimal()+team_theme+labs(y="Max B1",x="Hybrid Proportion",title="Max B1 vs Hybrid Proportion")
ggsave("maxB1vshybrid.png",width = 7, height = 5)

na.omit(B0B1)%>%
  ggplot(aes(x=On_Premises,y=max_B1))+geom_point(color="#F8766D")+geom_smooth(method = "lm", se=F, formula = y ~ x, alpha=0.1,color="#F8766D") + theme_minimal()+team_theme+labs(y="Max B1",x="On Premises Proportion",title="Max B1 vs On Premises Proportion")

ggsave("maxB1vsonpre.png",width = 7, height = 5)
```



# Plot for counties that have more than 50 deaths vs Plot for counties that have less than 50 deaths

```{r}
#Add a tag to counties to identify its death size.>=50: Large <50 small
death_prop_size<-death_prop%>%mutate(
  death_size=ifelse(CUMDEATHS>=50,"Large","Small")
)
death_prop_size$death_size<-as.factor(death_prop_size$death_size)
death_prop_size<-death_prop_size%>%select(COUNTY,death_size)
```


```{r}
# Max B1 vs Log of Population
na.omit(B0B1)%>%
  left_join(county_population,by="COUNTY")%>%
  left_join(death_prop_size,by="COUNTY")%>%
  ggplot(aes(x=log(POPULATION),y=max_B1,group=major_teaching,color=major_teaching))+geom_point(aes(size=avg_full_work_prob))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+team_theme+labs(y="Max B1",x="Log of Population",title="Max B1 vs Log of Population",size="Averaged \n%6hr+ Away Home")+
  facet_grid(.~death_size)
ggsave("facet_maxB1vslogpop.png",width = 7, height = 5)

# Max B1 vs Log of deaths during fall semester
na.omit(B0B1)%>%
  left_join(death_semester1,by="COUNTY")%>%
  left_join(death_prop_size,by="COUNTY")%>%
  ggplot(aes(x=log(death_semester),y=max_B1,group=major_teaching,color=major_teaching))+geom_point(aes(size=avg_full_work_prob))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+labs(y="Max B1",x="Log of Deaths during fall semester",title="Max B1 vs Log of Deaths during fall semester",size="Averaged \n%6hr+ Away Home")+
  facet_grid(.~death_size)
ggsave("facet_maxB1vslogdeathsem.png",width = 7, height = 5)

# Max B1 vs Log of deaths rate during fall semester
na.omit(B0B1)%>%
  left_join(death_semester1,by="COUNTY")%>%
  left_join(county_population,by="COUNTY")%>%
  left_join(death_prop_size,by="COUNTY")%>%
  mutate(log_death_rate=log(death_semester/POPULATION))%>%
  ggplot(aes(x=log_death_rate,y=max_B1,group=major_teaching,color=major_teaching))+geom_point(aes(size=avg_full_work_prob))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+labs(y="Max B1",x="Log of Death Rate",title="Max B1 vs Log of Death Rate during fall semester-log(death/population)",size="Averaged \n%6hr+ Away Home")+
  facet_grid(.~death_size)
ggsave("facet_maxB1vslogdeathrate1.png",width = 7, height = 5)
```

 

```{r}
# Max B1 vs Log of cumulative Deaths - the cut date(end of semester)
na.omit(B0B1)%>%
  left_join(death_semester1,by="COUNTY")%>%
  left_join(death_prop_size,by="COUNTY")%>%
  ggplot(aes(x=log(end),y=max_B1,group=major_teaching,color=major_teaching))+geom_point(aes(size=avg_full_work_prob))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+labs(y="Max B1",x="Log of total deaths",title="Max B1 vs Log of total deaths(Until 2020-12-15)",size="Averaged \n%6hr+ Away Home")+facet_grid(.~death_size)
ggsave("facet_maxB1vslogdeath.png",width = 7, height = 5)
```


```{r}
# Max B1 vs Log of cumulative Deaths - the cut date(2021-02-22)
death_total<-death_prop%>%select(COUNTY, CUMDEATHS)
na.omit(B0B1)%>%
  left_join(death_total,by="COUNTY")%>%
  left_join(death_prop_size,by="COUNTY")%>%
  ggplot(aes(x=log(CUMDEATHS),y=max_B1,group=major_teaching,color=major_teaching))+geom_point(aes(size=avg_full_work_prob))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+labs(y="Max B1",x="Log of total deaths",title="Max B1 vs Log of total deaths(Until 2021-02-22)",size="Averaged \n%6hr+ Away Home")+facet_grid(.~death_size)
ggsave("facet_maxB1vslogdeath2.png",width = 7, height = 5)

na.omit(B0B1)%>%
  left_join(mobility_semester,by="COUNTY")%>%
  left_join(county_population,by="COUNTY")%>%
  left_join(death_prop_size,by="COUNTY")%>%
  ggplot(aes(x=total_mobility,y=max_B1,group=major_teaching,color=major_teaching))+geom_point(aes(size=log(POPULATION)))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+labs(y="Max B1",x="Average mobility proportion during fall semester",title="Max B1 vs Average full mobility proportion during fall semester \nAveraged %6hr+ Away Home")+facet_grid(.~death_size)
ggsave("facet_maxB1vsavgmob.png",width = 7, height = 5)
```



# Plot for metro counties vs Plot for non-metro counties

```{r}
metro.definiotion<-ohio_profile%>%select(County,Metropolitan.Status)
names(metro.definiotion)[1]<-"COUNTY"
```


```{r}
# Max B1 vs Log of Population
na.omit(B0B1)%>%
  left_join(county_population,by="COUNTY")%>%
  left_join(metro.definiotion,by="COUNTY")%>%
  ggplot(aes(x=log(POPULATION),y=max_B1,group=major_teaching,color=major_teaching))+geom_point(aes(size=avg_full_work_prob))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+team_theme+labs(y="Max B1",x="Log of Population",title="Max B1 vs Log of Population",size="Averaged \n%6hr+ Away Home")+
  facet_grid(.~Metropolitan.Status)
ggsave("facet2_maxB1vslogpop.png",width = 7, height = 5)

# Max B1 vs Log of deaths during fall semester
na.omit(B0B1)%>%
  left_join(death_semester1,by="COUNTY")%>%
  left_join(metro.definiotion,by="COUNTY")%>%
  ggplot(aes(x=log(death_semester),y=max_B1,group=major_teaching,color=major_teaching))+geom_point(aes(size=avg_full_work_prob))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+labs(y="Max B1",x="Log of Deaths during fall semester",title="Max B1 vs Log of Deaths during fall semester",size="Averaged \n%6hr+ Away Home")+
  facet_grid(.~Metropolitan.Status)
ggsave("facet2_maxB1vslogdeathsem.png",width = 7, height = 5)

# Max B1 vs Log of deaths rate during fall semester
na.omit(B0B1)%>%
  left_join(death_semester1,by="COUNTY")%>%
  left_join(county_population,by="COUNTY")%>%
  left_join(metro.definiotion,by="COUNTY")%>%
  mutate(log_death_rate=log(death_semester/POPULATION))%>%
  ggplot(aes(x=log_death_rate,y=max_B1,group=major_teaching,color=major_teaching))+geom_point(aes(size=avg_full_work_prob))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+labs(y="Max B1",x="Log of Death Rate",title="Max B1 vs Log of Death Rate during fall semester-log(death/population)",size="Averaged \n%6hr+ Away Home")+
  facet_grid(.~Metropolitan.Status)
ggsave("facet2_maxB1vslogdeathrate1.png",width = 7, height = 5)
```

 

```{r}
# Max B1 vs Log of cumulative Deaths - the cut date(end of semester)
na.omit(B0B1)%>%
  left_join(death_semester1,by="COUNTY")%>%
  left_join(metro.definiotion,by="COUNTY")%>%
  ggplot(aes(x=log(end),y=max_B1,group=major_teaching,color=major_teaching))+geom_point(aes(size=avg_full_work_prob))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+labs(y="Max B1",x="Log of total deaths",title="Max B1 vs Log of total deaths(Until 2020-12-15)",size="Averaged \n%6hr+ Away Home")+facet_grid(.~Metropolitan.Status)
ggsave("facet2_maxB1vslogdeath.png",width = 7, height = 5)
```


```{r}
# Max B1 vs Log of cumulative Deaths - the cut date(2021-02-22)
death_total<-death_prop%>%select(COUNTY, CUMDEATHS)
na.omit(B0B1)%>%
  left_join(death_total,by="COUNTY")%>%
  left_join(metro.definiotion,by="COUNTY")%>%
  ggplot(aes(x=log(CUMDEATHS),y=max_B1,group=major_teaching,color=major_teaching))+geom_point(aes(size=avg_full_work_prob))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+labs(y="Max B1",x="Log of total deaths",title="Max B1 vs Log of total deaths(Until 2021-02-22)",size="Averaged \n%6hr+ Away Home")+facet_grid(.~Metropolitan.Status)
ggsave("facet2_maxB1vslogdeath2.png",width = 7, height = 5)

na.omit(B0B1)%>%
  left_join(mobility_semester,by="COUNTY")%>%
  left_join(county_population,by="COUNTY")%>%
  left_join(metro.definiotion,by="COUNTY")%>%
  ggplot(aes(x=total_mobility,y=max_B1,group=major_teaching,color=major_teaching))+geom_point(aes(size=log(POPULATION)))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+labs(y="Max B1",x="Average mobility proportion during fall semester",title="Max B1 vs Average full mobility proportion during fall semester \nAveraged %6hr+ Away Home")+facet_grid(.~Metropolitan.Status)
ggsave("facet2_maxB1vsavgmob.png",width = 7, height = 5)
```

# Boxplot of Max B1 for different teaching method.

```{r}
na.omit(B0B1)%>%
  ggplot(aes(x=major_teaching,y=max_B1,group=major_teaching,color=major_teaching))+geom_boxplot()+theme_minimal()+team_theme+
  labs(y="Max B1",x="Teaching Method",title="Boxplot of Max B1 vs Teaching Method")
ggsave("boxplotmaxB1vstmethod.png",width = 7, height = 5)
```

# Plot of population density
```{r}
na.omit(B0B1)%>%
  ggplot(aes(x=log(Population.density),y=max_B1,group=major_teaching,color=major_teaching))+geom_point(aes(size=avg_full_work_prob))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+labs(y="Max B1",x="Log Population Density",title="Max B1 vs Log Population Density",size="Averaged %6hr+ Away Home",color="Major Teaching")

ggsave("maxB1vsloglopden.png",width = 7, height = 5)
```


# Covid deaths during fall semester and death proportion during fall semester

```{r}
library(ggrepel)
library(cowplot)
library(sp)
ohio_map <- map_data("county") %>%subset(region=="ohio")%>%
  mutate(county=toupper(subregion))%>%select(long,lat,county,group)
getLabelPoint <- # Returns a county-named list of label points
function(county) {Polygon(county[c('long', 'lat')])@labpt}
centroids = by(ohio_map, ohio_map$county, getLabelPoint)     # Returns list
centroids2 <- do.call("rbind.data.frame", centroids)  # Convert to Data Frame
centroids2$county = str_to_title(rownames(centroids))
names(centroids2) <- c('clong', 'clat', "county")                 # Appropriate Header
```

```{r}
death_prop%>%
  left_join(ohio_map,by=c("COUNTY"='county'))%>%
  ggplot() + geom_polygon(aes(x = long, y = lat, group=group,fill = CUMDEATHS), color = "gray") + coord_fixed(1.3) + theme_map() +
scale_fill_distiller(palette = "OrRd",direction = 1)+labs(fill='Cumulative Deaths \nuntil 2021-02-22')+theme(legend.text = element_text(size=12),legend.title = element_text(size=12),legend.position = "bottom",legend.key.size = unit(2,"lines"))
ggsave("cumdeath.png",width = 5, height = 5)

death_prop%>%
  left_join(ohio_map,by=c("COUNTY"='county'))%>%
  ggplot() + geom_polygon(aes(x = long, y = lat, group=group,fill = death_per_1000), color = "gray") + coord_fixed(1.3) + theme_map() +
scale_fill_distiller(palette = "OrRd",direction = 1)+labs(fill='Deaths per 1000 people \nuntil 2021-02-22')+theme(legend.text = element_text(size=12),legend.title = element_text(size=12),legend.position = "bottom",legend.key.size = unit(2,"lines"))
ggsave("deathprop.png",width = 5, height = 5)
```

# 3 colored boxplots for response variable (total number of deaths before semester)/pop

```{r}
na.omit(death_semester1%>%
  left_join(B0B1%>%select(COUNTY,major_teaching),by="COUNTY")%>%
  left_join(death_prop%>%select(COUNTY,POPULATION),by="COUNTY")%>%
  mutate(d_inci_before_sem=start/POPULATION*1000))%>%
  ggplot(aes(x=major_teaching,y=d_inci_before_sem,group=major_teaching,color=major_teaching))+geom_boxplot()+theme_minimal()+team_theme+
  labs(y="Death Incidence",x="Teaching Method",title="Boxplot of Death Incidence vs Teaching Method \n-(total number of deaths before semester)/pop*1000",color="Teaching Method")
ggsave("boxplotmaxB1vstmethod_bfs.png",width = 7, height = 5)
```


# 3 colored boxplots for response variable (total number of deaths during semester)/pop

```{r}
na.omit(death_semester1%>%
  left_join(B0B1%>%select(COUNTY,major_teaching),by="COUNTY")%>%
  left_join(death_prop%>%select(COUNTY,POPULATION),by="COUNTY")%>%
  mutate(d_inci_before_sem=death_semester/POPULATION*1000))%>%
  ggplot(aes(x=major_teaching,y=d_inci_before_sem,group=major_teaching,color=major_teaching))+geom_boxplot()+theme_minimal()+team_theme+
  labs(y="Death Incidence",x="Teaching Method",title="Boxplot of Death Incidence vs Teaching Method \n-(total number of deaths during semester)/pop*1000",color="Teaching Method")
ggsave("boxplotmaxB1vstmethod_ds.png",width = 7, height = 5)
```

# Mobility map

# Cumulative Deaths for micro counties

```{r}
micor<-cases%>%
  mutate(cum_inci=CUMDEATHS/POPULATION*1000)%>%
  left_join(ohio_profile%>%
              select(County,NCHS.Urban.Rural.Status),
            by=c("COUNTY"="County"))%>%
  filter(NCHS.Urban.Rural.Status=="Micropolitan")%>%
  left_join(B0B1%>%select(COUNTY,major_teaching),by="COUNTY")

micor%>%
  group_by(major_teaching,DATE)%>%
  summarise(total_death=sum(CUMDEATHS),
            total_pop=sum(POPULATION))%>%
  mutate(cum_inci=total_death/total_pop*1000)%>%
    ggplot(aes(x=DATE,y=cum_inci,group=major_teaching,color=major_teaching))+geom_line()+theme_minimal()+team_theme+labs(y="Cumulative Death Incidence",x="Date",title="Cumulative Death Incidence for Micropolitan Counties\n-(total number of deaths)/pop*1000",color="Teaching Method")+
  annotate("rect", xmin=as.Date("2020/08/18"), xmax = as.Date("2020/12/15"), ymin = -Inf, ymax = Inf, fill = "orange", alpha = 0.2)+geom_vline(xintercept = as.Date("2020-11-12"), lty = 2)+geom_vline(xintercept = as.Date("2020-11-20"), lty = 2)
ggsave("cum_incidence_micro.png",width = 7, height = 5)
```



# Overlaid plots of cumulative deaths for micro counties

```{r}
micor%>%
  group_by(COUNTY,major_teaching,DATE)%>%
  summarise(total_death=sum(CUMDEATHS),
            total_pop=sum(POPULATION))%>%
  mutate(cum_inci=total_death/total_pop*1000)%>%
    ggplot(aes(x=DATE,y=cum_inci,group=COUNTY,color=major_teaching))+geom_line()+theme_minimal()+team_theme+labs(y="Cumulative Death Incidence",x="Date",title="Cumulative Death Incidence for Micropolitan Counties\n(Group by 22 Counties, Marked by Teaching Method)\n-(total number of deaths)/pop*1000",color="Teaching Method")+
  annotate("rect", xmin=as.Date("2020/08/18"), xmax = as.Date("2020/12/15"), ymin = -Inf, ymax = Inf, fill = "orange", alpha = 0.2)
ggsave("cum_incidence_micro2.png",width = 7, height = 5)
```




```{r}
ratio <- cases%>%
  left_join(wide_teaching_enroll,by=c("COUNTY"="county"))%>%
  drop_na(major_teaching)%>%
  group_by(major_teaching,DATE)%>%
  summarise(ratios = sum(CUMDEATHS)/sum(CUMCONFIRMED))

ratio%>%
  ggplot(aes(x=DATE,y=ratios,color=major_teaching))+geom_line(size=1)+
  geom_rect(data = ratio[1,],
            aes(xmin=as.Date("2020/08/18"), xmax=as.Date("2020/12/15"),
                ymin=-Inf,ymax=Inf),
            color = NA,alpha=0.2, show.legend = F, fill = "orange") + 
  geom_vline(xintercept = as.Date("2020/11/24"), lty = 2) +
  annotate("text",label = as.Date("2020/11/24"),
           x = as.Date("2020/11/24"), y = .07, hjust = 1.1)+ 
  labs(x = "Date", y = "Cumulative deaths/cases ratio", 
       color = "Majority Teaching Method",
       subtitle = "Yellow area represents the fall semester (08/18 - 12/15)") + 
  theme(legend.position = "bottom")+team_theme

cases%>%
  left_join(wide_teaching_enroll,by=c("COUNTY"="county"))%>%
  drop_na(major_teaching)%>%
  group_by(major_teaching,DATE)%>%
  summarise(cum=sum(CUMCONFIRMED*major_teaching_prop))%>%
  ggplot(aes(x=DATE,y=log(cum),color=major_teaching))+geom_line(size=1)+
    geom_vline(xintercept = as.Date("2020/11/24"), lty = 2) +
  annotate("text",label = as.Date("2020/11/24"),
           x = as.Date("2020/11/24"), y = .07, hjust = 1.1)+ 
  labs(x = "Date", y = "Logged Cumulative Cases", 
       color = "Majority Teaching Method",
       subtitle = "Yellow area represents the fall semester (08/18 - 12/15)") + 
  theme(legend.position = "bottom")+
  theme(axis.line = element_line(colour = "black"),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.border = element_blank(),
        panel.background = element_blank())+
  theme(axis.text = element_text(size=15),axis.title=element_text(size=15),legend.text = element_text(size=13))
```

## Confounders

```{r}
library(ggpubr)
library(PMCMRplus)
require(DescTools)

teaching_profile <- ohio_profile%>%
  inner_join(wide_teaching_enroll,by=c("County"="county"))
teaching_profile$major_teaching <- factor(teaching_profile$major_teaching,levels = c("On Premises","Hybrid","Online Only"))

profile_major_teaching.aov <- aov(Percent.uninsured ~ major_teaching,data = teaching_profile)
summary(profile_major_teaching.aov)

# p-value of .012
stat.test <- PostHocTest(profile_major_teaching.aov, method = "duncan")$major_teaching %>%
  as.data.frame()%>%
  rownames_to_column("group") %>%
  separate(group,"-", into = c("group1","group2")) %>%
  mutate(pval = round(pval,3),
         p = case_when(pval <= .01~ "**",
                       pval <= .05 ~ "*",
                       TRUE ~ "NS"))%>%
  select(group1, group2, pval, p)



teaching_profile%>%
  ggplot(aes(x=major_teaching,y=Percent.uninsured))+geom_boxplot(aes(fill=major_teaching),width=0.6)+
  stat_compare_means(method = "anova",size=6,label.y.npc=0.85)+ 
  stat_pvalue_manual(stat.test, label = "p",y.position = 1, step.increase = 0.15,size = 6,bracket.nudge.y = 8)+
  labs(title="Percent of Uninsured Population by Teaching Method",x="Majority Teaching Method",y="Percent of uninsured population",caption = "Post-hoc pairwise testing with Duncan Method")+
  team_theme+theme(legend.position = "")

ggsave("p3.png",width = 6.5,height = 6)
```



```{r}

senior_major_teaching.aov <- aov(Percent.Population.65..yrs~ major_teaching,data = teaching_profile)

summary(senior_major_teaching.aov)

# p-value of .012
stat.test <- PostHocTest(senior_major_teaching.aov, method = "duncan")$major_teaching %>%
  as.data.frame()%>%
  rownames_to_column("group") %>%
  separate(group,"-", into = c("group1","group2")) %>%
  mutate(pval = round(pval,3),
         p = case_when(pval <= .01~ "**",
                       pval <= .05 ~ "*",
                       TRUE ~ "NS"))%>%
  select(group1, group2, pval, p)

teaching_profile%>%
  ggplot(aes(x=major_teaching,y=Percent.Population.65..yrs))+geom_boxplot(aes(fill=major_teaching),width=0.6)+
  stat_compare_means(method = "anova",size=6,label.y.npc=0.95)+ 
  stat_pvalue_manual(stat.test, label = "p",y.position = 1, step.increase = 0.15,size = 5,bracket.nudge.y = 2)+
  labs(title="Percent of 65years+ Population by Teaching Method",x="Major Teaching Method",y="Percent of 65+yrs population",fill="Majority Teaching Method",caption = "Post-hoc pairwise testing with Duncan Method")+team_theme+theme(legend.position = "")

ggsave("p3.png",width = 6.5,height = 6)
```
