==> RawCodes/step13_Update_bplots.Rmd <==
---
title: "Update B plots"
author: "Ziyan Zhu"
date: "4/24/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Notes:

The 1st derivative of log new deaths+1 we observed on date T corresponds to the death growth rate B happend on date T-24.

The mobility date has been shifted one week forward as well. Assuming mobility a week before date T change the infections on date T


# Load in the spline slopes

```{r message=FALSE, warning=FALSE, cache=TRUE}
#source("step2_data_wrangle.r") ## B(t) include in step2 as well

# Aggregate counties by teaching method
cases_slope_teach_agg <- cases_slope_teach %>%
  drop_na(major_teaching)%>%
  group_by(DATE, major_teaching) %>%
  summarise(total_new_deaths = sum(rev_NEWDEATHS), .groups = "drop") %>%
  mutate(log_new_deaths = log(total_new_deaths + 1)) %>%
  group_by(major_teaching) %>%
  mutate(smooth.spline = smooth.spline(DATE,log_new_deaths,df = 398/28)$y,
         B = predict(smooth.spline(DATE,log_new_deaths,df = 398/28),deriv = 1)$y,B2 = predict(smooth.spline(DATE,log_new_deaths,df = 398/28),deriv = 2)$y)

## plot theme
team_theme <- theme(axis.line = element_line(colour = "black"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_blank(),
    panel.background = element_blank(),legend.position = "bottom",legend.box="vertical", legend.margin=margin())+theme(legend.text = element_text(size=12),legend.title = element_text(size=12),axis.text = element_text(size=13),title=element_text(size=13))
```


## B plots

```{r}
ggplot(cases_slope_teach_agg, aes(x = DATE, color = major_teaching)) +
  geom_point(aes(y = log_new_deaths), alpha = .3)+ 
  geom_line(aes(y = smooth.spline), size = 1) +
  geom_rect(data = cases_slope_teach_agg[1,],
            aes(xmin=as.Date("2020/08/18"), xmax=as.Date("2020/12/15"),
                ymin=-Inf,ymax=Inf),
            color = NA,alpha=0.2, show.legend = F, fill = "orange") + 
  theme_bw() + 
  labs(x = "Date", y = "Log ( New Deaths + 1 )", 
       color = "Majority Teaching Method",
       caption = "Smoothing window set to every 4 weeks",
       subtitle = "Yellow area represents the fall semester (08/18 - 12/15)") + 
  geom_vline(xintercept = week3_after_start, lty = 2) + 
  annotate("text",label = "3 Weeks After",
           x = week3_after_start, y = 4, hjust = 1.1)+ 
  geom_vline(xintercept = as.Date("2020/08/18")+42, lty = 2) + 
  annotate("text",label = "6 Weeks After",
           x = as.Date("2020/08/18")+62, y = 4.7, hjust = 1.3)+ 
  theme(legend.position = "bottom")+
  theme(axis.line = element_line(colour = "black"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_blank(),
    panel.background = element_blank())+
  theme(axis.text = element_text(size=15),axis.title=element_text(size=15),legend.text = element_text(size=13))

#ggsave("log_death_time_series.png",width = 7, height = 5)


week3_after_start <- as.Date("2020/08/18") + 21

ggplot(cases_slope_teach_agg, aes(x = DATE, color = major_teaching)) + 
  geom_line(aes(y = B), size = 1) +
  geom_rect(data = cases_slope_teach_agg[1,],
            aes(xmin=as.Date("2020/08/18"), xmax=as.Date("2020/12/15"),
                ymin=-Inf,ymax=Inf),
            color = NA,alpha=0.2, show.legend = F, fill = "orange") +
  geom_vline(xintercept = week3_after_start, lty = 2) + 
  annotate("text",label = "3 Weeks After",
           x = week3_after_start, y = .05, hjust = 1.1)+ 
  geom_vline(xintercept = as.Date("2020/08/18")+42, lty = 2) + 
  annotate("text",label = "6 Weeks After",
           x = as.Date("2020/08/18")+130, y = .06, hjust = 1.3)+ 
  labs(x = "Date", y = "Exponential Growth Coefficient", 
       color = "Majority Teaching Method",
       caption = "Smoothing window set to every 4 weeks",
       subtitle = "Yellow area represents the fall semester (08/18 - 12/15)") + 
  theme(legend.position = "bottom")+
  theme(axis.line = element_line(colour = "black"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_blank(),
    panel.background = element_blank())+
  theme(axis.text = element_text(size=15),axis.title=element_text(size=15),legend.text = element_text(size=13))
#ggsave("egc_time_series.png",width = 7, height = 5)
```



```{r}
## derivative of B, B2
ggplot(cases_slope_teach_agg, aes(x = DATE, color = major_teaching)) + 
  geom_line(aes(y = B2), size = 1) +
  geom_rect(data = cases_slope_teach_agg[1,],
            aes(xmin=as.Date("2020/08/18"), xmax=as.Date("2020/12/15"),
                ymin=-Inf,ymax=Inf),
            color = NA,alpha=0.2, show.legend = F, fill = "orange") +
  geom_vline(xintercept = week3_after_start, lty = 2) + 
  annotate("text",label = "3 Weeks After",
           x = week3_after_start, y = .002, hjust = 1.1)+ 
  geom_vline(xintercept = as.Date("2020/08/18")+42, lty = 2) + 
  annotate("text",label = "6 Weeks After",
           x = as.Date("2020/08/18")+130, y = .003, hjust = 1.3)+ 
  labs(x = "Date", y = "2nd Derivative of Smoothing Splines", 
       color = "Majority Teaching Method",
       caption = "Smoothing window set to every 4 weeks",
       subtitle = "Yellow area represents the fall semester (08/18 - 12/15)") + 
  theme(legend.position = "bottom")+
  theme(axis.line = element_line(colour = "black"),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.border = element_blank(),
        panel.background = element_blank())+
  theme(axis.text = element_text(size=15),axis.title=element_text(size=15),legend.text = element_text(size=13))
```

# Max B1 vs Average B1 in fall

```{r}
# maxB1 vs average B1
na.omit(B0B1)%>%
  ggplot(aes(x=avg_B1,y=max_B1,group=major_teaching,color=major_teaching))+geom_point(aes(size=On_Premises,alpha=On_Premises))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+labs(color="Majority Teaching Method",size="Percent of On-premises",title="Max B1 vs Average B1",alpha="Percent of On-premises",y="Max B1 in Fall", x= "Averaged B1 in Fall")+
  team_theme+team_theme+geom_vline(xintercept = 0, lty = 2)
#ggsave("maxB1vsavgB1.png",width = 7, height = 5)
```



## Average mobility during first 6 weeks of Fall

```{r}
# average full time work mobility in 6 weeks window
avg_mobility <- case_mobility%>%
  left_join(major_reopening,by=c("COUNTY"))%>%
  group_by(COUNTY)%>%
  filter(DATE >= as.Date(major_opendate) + 21 & DATE <= as.Date(major_opendate) + 63)%>%
  summarise(avg_full_work_prob = mean(full_work_prop_7d))

# made major teaching method prop
wide_teaching_enroll <- wide_teaching_enroll%>%
  mutate(major_teaching_prop = case_when(
    major_teaching=="Online Only" ~ Online_Only,
    major_teaching=="Hybrid" ~ Hybrid,
    major_teaching=="On Premises" ~On_Premises,
    TRUE~ 0
  ))
```


### B1 vs Population

```{r}
# Max B1 vs Population
na.omit(B0B1)%>%
  drop_na(major_teaching)%>%
  ggplot(aes(x=POPULATION,y=max_B1,group=major_teaching,color=major_teaching))+geom_point(aes(size=avg_full_work_prob,alpha=avg_full_work_prob))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+team_theme+geom_vline(xintercept = 0, lty = 2)+labs(y="Max B1",x="Population",title="Max B1 vs Population",color="Majority Teaching Method",size = "Averaged \n%6hr+ Away Home",alpha= "Averaged \n%6hr+ Away Home" )

# Max B1 vs Log of Population
na.omit(B0B1)%>%
  drop_na(major_teaching)%>%
  ggplot(aes(x=log(POPULATION),y=max_B1,group=major_teaching,color=major_teaching))+geom_point(aes(size=avg_full_work_prob,alpha=avg_full_work_prob))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+team_theme+labs(y="Max B1",x="Log of Population",title="Max B1 vs Log of Population",color="Majority Teaching Method",size = "Averaged \n%6hr+ Away Home",alpha= "Averaged \n%6hr+ Away Home" )

# Max B1 vs Log of Population Density
na.omit(B0B1)%>%
  drop_na(major_teaching)%>%
  ggplot(aes(x=log(Population.density),y=max_B1,group=major_teaching,color=major_teaching))+geom_point(aes(size=avg_full_work_prob,alpha=avg_full_work_prob))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+team_theme+labs(y="Max B1",x="Log of Population Density",title="Max B1 vs Log of Population Density",color="Majority Teaching Method",size = "Averaged \n%6hr+ Away Home",alpha= "Averaged \n%6hr+ Away Home" )
```



# Mobility

## MaxB1 vs Mobility

```{r}
# Max B1 vs average Mobility proportion
mobility_semester<-case_mobility%>%
  filter(DATE>=as.Date("2020-08-18") & DATE<=as.Date("2020-12-15"))%>%
  select(COUNTY,full_work_prop_7d)%>%group_by(COUNTY)%>%
  summarise(total_mobility=mean(full_work_prop_7d))

na.omit(B0B1)%>%
  left_join(mobility_semester,by="COUNTY")%>%
  ggplot(aes(x=total_mobility,y=max_B1,group=major_teaching,color=major_teaching))+geom_point(aes(size=log(POPULATION),alpha=log(POPULATION)))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+labs(y="Max B1",x="Average mobility proportion during fall semester",title="Max B1 vs Average full mobility proportion during fall semester \nAveraged %6hr+ Away Home",size="Log of Population",alpha="Log of Population",color="Majority Teaching Method")
```

```{r}
# max B1 vs RURAL
B0B1%>%
  drop_na(major_teaching)%>%
  ggplot(aes(y=NCHS.Urban.Rural.Status,x=max_B1,fill=major_teaching))+facet_grid(~major_teaching)+geom_boxplot()+labs(color="Majority Teaching Method",size="Averaged \n%6hr+ Away Home",title="max B1 v.s. Urban Rural Status",alpha="Size of Majority",x="Max B1", y= "Urban Rural Status")+team_theme
```


```{r}

na.omit(B0B1)%>%
  filter(NCHS.Urban.Rural.Status=="Micropolitan")%>%
  drop_na(major_teaching)%>%
  ggplot(aes(x=log(Population.density),y=max_B1,group=major_teaching,color=major_teaching))+geom_point(aes(size=avg_full_work_prob,alpha=avg_full_work_prob))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+team_theme+labs(y="Max B1",x="Log of Population Density",title="Max B1 in Fall v.s. Log of Population Density\nOnly Micropolitan Counties",color="Majority Teaching Method",size = "Averaged \n%6hr+ Away Home",alpha= "Averaged \n%6hr+ Away Home" )

na.omit(B0B1)%>%
  filter(NCHS.Urban.Rural.Status=="Micropolitan")%>%
  left_join(mobility_semester,by="COUNTY")%>%
  ggplot(aes(x=total_mobility,y=max_B1,group=major_teaching,color=major_teaching))+geom_point(aes(size=log(POPULATION),alpha=log(POPULATION)))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+labs(y="Max B1",x="Average mobility proportion during fall semester",title="Max B1 v.s. Average full mobility proportion during Fall\nOnly Micropolitan Counties",size="Log of Population",alpha="Log of Population",color="Majority Teaching Method")
ggsave("p2.png",width = 6.5,height = 6)
```



# Appendix in cool plots

## Max B1-Log population Regression Part

```{r}
fit1 <- lm(max_B1~log(POPULATION),data=B0B1,weights = B0B1$Online_Only)
summary(fit1)

fit2 <- lm(max_B1~log(POPULATION),data=B0B1,weights = B0B1$On_Premises)
summary(fit2)

fit3 <- lm(max_B1~log(POPULATION),data=B0B1,weights = B0B1$Hybrid)
summary(fit3)
```
