==> RawCodes/step14_v2.rmd <==
---
title: "B1 minus B0"
author: "Ziyan Zhu"
date: "4/24/2021"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Notes:

The 1st derivative of log new deaths+1 we observed on date T corresponds to the death growth rate B happend on date T-24.

The mobility date has been shifted one week forward as well. Assuming mobility a week before date T change the infections on date T


# Load in the spline slopes

```{r message=FALSE, warning=FALSE, cache=TRUE}
source("step2_data_wrangle.r") ## B(t) include in step2 as well
## plot theme
team_theme <- theme(axis.line = element_line(colour = "black"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_blank(),
    panel.background = element_blank(),legend.position = "bottom",legend.box="vertical", legend.margin=margin())+theme(legend.text = element_text(size=12),legend.title = element_text(size=12),axis.text = element_text(size=13),title=element_text(size=13))
```

# compute the differences in B(3weeks after start) and B(6weeks after start)

```{r}
B3w <- cases_slope_teach%>%
  filter(DATE==as.Date("2020-08-18")+21)%>%
  drop_na(major_teaching)
B6w <- cases_slope_teach%>%
  filter(DATE==as.Date("2020-08-18")+42)%>%
  drop_na(major_teaching)%>%
  rename(new.slope6w=new.slope)
avg_mobi_3w <- case_mobility%>%
  left_join(major_reopening,by=c("COUNTY"))%>%
  group_by(COUNTY)%>%
  filter(DATE >= as.Date(major_opendate) + 21 & DATE <= as.Date(major_opendate) + 42)%>%
  summarise(avg3w_full_work_prob = mean(full_work_prop_7d))
B4w <- cases_slope_teach%>%
  filter(DATE==as.Date("2020-08-18")+28)%>%
  drop_na(major_teaching)
avg_mobi_4w <- case_mobility%>%
  left_join(major_reopening,by=c("COUNTY"))%>%
  group_by(COUNTY)%>%
  filter(DATE >= as.Date(major_opendate) + 28 & DATE <= as.Date(major_opendate) + 42)%>%
  summarise(avg4w_full_work_prob = mean(full_work_prop_7d))
B_diff1 <- B6w[,c(1:9,13,20)]%>%
  left_join(B3w%>%select(COUNTY,new.slope),by="COUNTY")%>%
  mutate(new.slope.diff = new.slope6w-new.slope)%>%
  left_join(avg_mobi_3w,by="COUNTY")%>%rename(avg_mob=avg3w_full_work_prob)
B_diff2 <- B6w[,c(1:9,13,20)]%>%
  left_join(B4w%>%select(COUNTY,new.slope),by="COUNTY")%>%
  mutate(new.slope.diff = new.slope6w-new.slope)%>%
  left_join(avg_mobi_4w,by="COUNTY")%>%rename(avg_mob=avg4w_full_work_prob)
```



# Add B4w
```{r}
B_diff1<-B_diff1%>%mutate(type="6w - 3w")
B_diff2<-B_diff2%>%mutate(type="6w - 4w")
B_diff1$new.slope.diff<-B_diff1$new.slope.diff/3
B_diff2$new.slope.diff<-B_diff2$new.slope.diff/2
```

```{r}
B_new_diff<-rbind(B_diff1,B_diff2)
```

## Reproduce cool plots - All

```{r}
# Population Density
B_new_diff%>%
  drop_na(major_teaching)%>%
  ggplot(aes(x=log(Population.density),y=new.slope.diff,group=major_teaching,color=major_teaching))+geom_point(aes(size=avg_mob,alpha=avg_mob))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+team_theme+labs(y="Avg week B1-B0",x="Log of Population Density",title="Avg week B1-B0 vs Log of Population Density",color="Majority Teaching Method",size = "Averaged %6hr+ Away Home",alpha= "Averaged %6hr+ Away Home" )+facet_grid(.~type)

# Avg mobility
B_new_diff%>%
  drop_na(major_teaching)%>%
  ggplot(aes(x=avg_mob,y=new.slope.diff,group=major_teaching,color=major_teaching))+geom_point(aes(size=log(Population.density),alpha=log(Population.density)))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+labs(y="Avg week B1-B0",x="Average mobility proportion during fall semester",title="Avg week B1-B0 vs Average full mobility proportion\n Averaged %6hr+ Away Home",size="Log of Population Density",alpha="Log of Population Density",fill="Majority Teaching Method")+facet_grid(.~type)

# Urban Status
B_new_diff%>%
  drop_na(major_teaching)%>%
  ggplot(aes(y=NCHS.Urban.Rural.Status,x=new.slope.diff,fill=NCHS.Urban.Rural.Status))+geom_boxplot()+team_theme+theme(legend.position = "")+labs(x="Avg week B1-B0",y="Rural-urban Status",title="Distribution of (B1 - B0)/N of weeks")+
  scale_fill_brewer(palette = "Reds",direction = -1)+facet_grid(.~type)
B_new_diff%>%
  drop_na(major_teaching)%>%
  ggplot(aes(y=Metropolitan.Status,x=new.slope.diff,fill=Metropolitan.Status))+geom_boxplot()+team_theme+theme(legend.position = "")+labs(x="Avg week B1-B0",y="Metropolitan Status",title="Distribution of (B1 - B0)/N of weeks")+
  scale_fill_brewer(palette = "Reds",direction = -1) +facet_grid(.~type)
```

## Reproduce cool plots - Micropolitan

```{r}
# Filter only the micro counties
B_new_diff_micro <- B_new_diff%>%
  drop_na(major_teaching)%>%
  filter(NCHS.Urban.Rural.Status=="Micropolitan")

# Population Density
B_new_diff_micro%>%
  ggplot(aes(x=log(Population.density),y=new.slope.diff,group=major_teaching,color=major_teaching))+geom_point(aes(size=avg_mob,alpha=avg_mob))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+team_theme+labs(y="B1-B0",x="Log of Population",title="B1-B0 vs Log of Population Density\nOnly Micropolitan Counties",color="Majority Teaching Method",size = "3w~6w\nAveraged %6hr+ Away Home",alpha= "3w~6w\nAveraged %6hr+ Away Home" )+facet_grid(.~type)

# Avg mobility
B_new_diff_micro%>%
  ggplot(aes(x=avg_mob,y=new.slope.diff,group=major_teaching,color=major_teaching))+geom_point(aes(size=log(Population.density),alpha=log(Population.density)))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+labs(y="B1-B0",x="Average mobility proportion during fall semester",title="B1-B0 vs Average full mobility proportion in 3w~6w\n Averaged %6hr+ Away Home\nOnly Micropolitan Counties",size="Log of Population Density",alpha="Log of Population Density",fill="Majority Teaching Method")+facet_grid(.~type)

```

# Look at 10 micropolitan counties

```{r}
B_10<-cases_slope_teach%>%
  filter(major_teaching=="On Premises"|major_teaching=="Online Only")%>%
  filter(NCHS.Urban.Rural.Status=="Micropolitan")
B_10%>%distinct(COUNTY,major_teaching)
```

## On Premises Micro county

```{r}
Sys.setlocale("LC_TIME", "English")
B_10%>%filter(major_teaching=="On Premises")%>%
  ggplot(aes(x=DATE,y=new.slope))+
  geom_line(color = "#F8766D",size = 1)+theme_minimal()+team_theme+
  facet_wrap(~ COUNTY, nrow = 4)+labs(y="B",x="Date",title="B with 24 days shifted across to time - On Premises Micropolitan Counties")+
  annotate("rect", xmin=as.Date("2020/08/18"), xmax = as.Date("2020/12/15"), ymin = -Inf, ymax = Inf, fill = "orange", alpha = 0.2)

ggsave("micro_onpremises.png",height = 7,width = 8)
```





```{r}
B_10%>%filter(major_teaching=="Online Only")%>%
  ggplot(aes(x=DATE,y=new.slope))+
  geom_line(color = "#619CFF",size = 1)+theme_minimal()+team_theme+
  facet_wrap(~ COUNTY, nrow = 2)+labs(y="B",x="Date",title="B with 24 days shifted across to time - Online Only Micropolitan Counties")+
  annotate("rect", xmin=as.Date("2020/08/18"), xmax = as.Date("2020/12/15"), ymin = -Inf, ymax = Inf, fill = "orange", alpha = 0.2)

ggsave("micro_online.png",height = 3.5,width = 8)
```
