==> RawCodes/step14_what_changes_B.Rmd <==
---
title: "B1 minus B0"
author: "Ziyan Zhu"
date: "4/24/2021"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Notes:

The 1st derivative of log new deaths+1 we observed on date T corresponds to the death growth rate B happend on date T-24.

The mobility date has been shifted one week forward as well. Assuming mobility a week before date T change the infections on date T


# Load in the spline slopes

```{r message=FALSE, warning=FALSE, cache=TRUE}
source("step2_data_wrangle.r") ## B(t) include in step2 as well

## plot theme
grid_theme <- theme(axis.line = element_line(colour = "black"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_blank(),
    panel.background = element_blank(),legend.position = "bottom",legend.box="vertical", legend.margin=margin(),panel.spacing = unit(1.5, "lines"),strip.text.x = element_text(
        size = 12, face = "bold.italic"))

team_theme <- grid_theme+theme(legend.text = element_text(size=13),legend.title = element_text(size=13),axis.text = element_text(size=13),title=element_text(size=14),axis.title.x.bottom = element_text(margin = margin(12, 10, 6, 10)),axis.title.y.left = element_text(size=14),axis.text.y.left = element_text(size = 14))+theme( plot.caption = element_text(face = "italic",size = 11.5))
```

# compute the differences in B(3weeks after start) and B(6weeks after start)

```{r}
B0w <- cases_slope_teach%>%
  filter(DATE==as.Date("2020-08-18"))%>%
  drop_na(major_teaching)%>%
  rename(new.slope0w=new.slope)


B3w <- cases_slope_teach%>%
  filter(DATE==as.Date("2020-08-18")+21)%>%
  drop_na(major_teaching)%>%
  rename(new.slope3w=new.slope)

B6w <- cases_slope_teach%>%
  filter(DATE==as.Date("2020-08-18")+42)%>%
  drop_na(major_teaching)%>%
  rename(new.slope6w=new.slope)

avg_mobi_0w3w <- case_mobility%>%
  left_join(major_reopening,by=c("COUNTY"))%>%
  group_by(COUNTY)%>%
  filter(DATE >= as.Date("2020-08-18")& DATE <as.Date("2020-08-18") + 21)%>%
  summarise(avg_full_work_prob = mean(full_work_prop_7d))

avg_mobi_3w6w <- case_mobility%>%
  left_join(major_reopening,by=c("COUNTY"))%>%
  group_by(COUNTY)%>%
  filter(DATE >= as.Date("2020-08-18")+ 21 & DATE <=as.Date("2020-08-18") + 42)%>%
  summarise(avg2_full_work_prob = mean(full_work_prop_7d))


B_diff <- B6w[,c(1:9,13,20)]%>%
  left_join(B3w%>%select(COUNTY,new.slope3w),by="COUNTY")%>%
  left_join(B0w%>%select(COUNTY,new.slope0w),by="COUNTY")%>%
  mutate(new.slope.diff = new.slope3w-new.slope0w,new.slope.diff2 = new.slope6w-new.slope3w)%>%
  left_join(avg_mobi_0w3w,by="COUNTY")%>%
  left_join(avg_mobi_3w6w,by="COUNTY")


B_diff$major_teaching <- factor(B_diff$major_teaching,levels = c("On Premises","Hybrid","Online Only"))


hist(B_diff$new.slope.diff)
hist(B_diff$new.slope.diff2)
```


```{r}
# pop density plot

B_diff%>%
  drop_na(major_teaching)%>%
  ggplot(aes(x=log(Population.density),y=new.slope.diff2,group=major_teaching,color=major_teaching))+geom_point(aes(size=avg2_full_work_prob,alpha=avg2_full_work_prob))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+team_theme+labs(y="B(6w) - B(3w)",x="Log of Population Density",title="B1-B0 vs Log of Population Density",color="Majority Teaching Method",size = "3w~6w\nAveraged %6hr+ Away Home",alpha= "3w~6w\nAveraged %6hr+ Away Home" )
ggsave("p1.png",width =7,height =  6.5)

# mobility plot

B_diff%>%
  drop_na(major_teaching)%>%
  ggplot(aes(x=avg2_full_work_prob,y=new.slope.diff2,group=major_teaching,color=major_teaching))+geom_point(aes(size=log(Population.density),alpha=log(Population.density)))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+labs(y="B(6w) - B(3w)",x="Average mobility proportion during fall semester",title="B1-B0 vs Average full mobility proportion in 3w~6w\n Averaged %6hr+ Away Home",size="Log of Population Density",alpha="Log of Population Density",color="Majority Teaching Method")
ggsave("p2.png",width = 7,height =  6.5)

```

```{r}
B_diff%>%
  drop_na(major_teaching)%>%
  ggplot(aes(y=NCHS.Urban.Rural.Status,x=new.slope.diff2,fill=NCHS.Urban.Rural.Status))+geom_boxplot()+team_theme+theme(legend.position = "")+labs(x="B(6w) - B(3w)",y="Rural-urban Status",title="Distribution of B(6 weeks after) minus B(3 weeks after)")+
  scale_fill_brewer(palette = "Reds",direction = -1)

B_diff%>%
  drop_na(major_teaching)%>%
  ggplot(aes(y=NCHS.Urban.Rural.Status,x=new.slope.diff2,fill=major_teaching))+geom_boxplot()+facet_grid(~major_teaching)+theme(legend.position = "")+labs(x="B(6w) - B(3w)",y="Rural-urban Status",title="Distribution of B(6 weeks after) minus B(3 weeks after)\nby Rural-urban Status",fill="Majority Teaching Method")+team_theme

ggsave("p2.png",width = 10,height = 5)

B_diff%>%
  drop_na(major_teaching)%>%
  ggplot(aes(y=Metropolitan.Status,x=new.slope.diff2,fill=Metropolitan.Status))+geom_boxplot()+team_theme+theme(legend.position = "")+labs(x="B(6w) - B(3w)",y="Metropolitan Status",title="Distribution of B(6 weeks after) minus B(3 weeks after)")+
  scale_fill_brewer(palette = "Reds",direction = -1) 

```

```{r}
B_diff_micro%>%filter(COUNTY=="CHAMPAIGN")
B_diff_micro%>%filter(COUNTY=="DEFIANCE")
B_diff_micro%>%filter(COUNTY=="MERCER")
```


## Redo the plots but only with Micro counties

```{r}
## Filter only the micro counties
B_diff_micro <- B_diff%>%
  drop_na(major_teaching)%>%
  filter(NCHS.Urban.Rural.Status=="Micropolitan") %>%
  mutate(acc = new.slope.diff2 - new.slope.diff)

B_diff_micro%>%
  ggplot(aes(x=avg_full_work_prob,y=new.slope.diff,group=major_teaching,color=major_teaching))+geom_point(aes(size=log(Population.density),alpha=log(Population.density)))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+labs(y="B(3w) - B(0w)",x="Average mobility proportion in 0w-3w",title="B1-B0 in 0w~3w v.s. Average full mobility proportion\nOnly Micropolitan Counties",size="Log of Population Density",alpha="Log of Population Density",fill="Majority Teaching Method",fill="Majority Teaching Method")+team_theme+geom_text(data =B_diff_micro%>%filter(major_teaching=="On Premises"),aes(label=COUNTY),color='black',size=3,hjust=1.1, vjust=0.3,size=6)
#ggsave("p1.png",width = 7.5,height = 6)


B_diff_micro%>%
  ggplot(aes(x=log(Population.density),y=new.slope.diff2,group=major_teaching,color=major_teaching))+geom_point(aes(size=avg2_full_work_prob,alpha=avg2_full_work_prob))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+team_theme+labs(y="B(6w) - B(3w)",x="Log of Population Density",title="B1-B0 in 3w~6w v.s. Log of Population Density\nOnly Micropolitan Counties",color="Majority Teaching Method",size = "3w~6w\nAveraged %6hr+ Away Home",alpha= "3w~6w\nAveraged %6hr+ Away Home" ,fill="Majority Teaching Method")+geom_text(data =B_diff_micro%>%filter(major_teaching=="On Premises"),aes(label=COUNTY),color='black',size=3,hjust=1.1, vjust=0.3)
#ggsave("p2.png",width = 7.5,height = 6)

```

```{r}
B_diff_micro%>%
  ggplot(aes(x=log(Population.density),y=new.slope.diff,group=major_teaching,color=major_teaching))+geom_point(aes(size=avg_full_work_prob,alpha=avg_full_work_prob))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+labs(y="B(3w) - B(0w)",x="Log of Population Density",title="B1-B0 in 0w~3w v.s. Log of Population Density\nOnly Micropolitan Counties",color="Majority Teaching Method",size = "0w~3w\nAveraged %6hr+ Away Home",alpha= "0w~3w\nAveraged %6hr+ Away Home" ,fill="Majority Teaching Method")+team_theme+geom_text(data =B_diff_micro%>%filter(major_teaching=="On Premises"),aes(label=COUNTY),color='black',size=3,hjust=1.1, vjust=0.3,size=6)

ggsave("p1.png",width = 7.5,height = 6)


B_diff_micro%>%
  ggplot(aes(x=avg2_full_work_prob,y=new.slope.diff2,group=major_teaching,color=major_teaching,label=COUNTY))+geom_point(aes(size=log(Population.density),alpha=log(Population.density)))+geom_smooth(method = "lm", se=F, formula = y ~ x,alpha=0.1)+theme_minimal()+team_theme+labs(y="B(6w) - B(3w)",x="Average mobility proportion in 3w-6w",title="B1-B0 in 3w~6w v.s. Average full mobility proportion \nOnly Micropolitan Counties",size="Log of Population Density",alpha="Log of Population Density",fill="Majority Teaching Method")+geom_text(data =B_diff_micro%>%filter(major_teaching=="On Premises"),aes(label=COUNTY),color='black',size=3,hjust=1.1, vjust=0.1)

ggsave("p2.png",width = 7.5,height = 6)
```

### B1-B0 dot plots for each group

```{r}
B_diff_micro%>%
  ggplot(aes(x=avg_full_work_prob,y=new.slope.diff,group=major_teaching,color=major_teaching))+geom_point(aes(size=log(Population.density),alpha=log(Population.density)))+facet_grid(~major_teaching)+labs(y="B(3w) - B(0w)",x="Average mobility proportion in 0w-3w",title="B1(3w)-B0(0w) in Only Micropolitan Counties",size="Log of Population Density",color="Majority Teaching Method",alpha="Log of Population Density",fill="Majority Teaching Method")+grid_theme+theme(legend.text = element_text(size=13),legend.title = element_text(size=13),axis.text = element_text(size=11),title=element_text(size=14),strip.text.x = element_text(
        size = 12, face = "bold.italic"
        ))

ggsave("p1.png",width = 10,height = 4)


B_diff_micro%>%
  ggplot(aes(x=avg2_full_work_prob,y=new.slope.diff2,group=major_teaching,color=major_teaching))+geom_point(aes(size=log(Population.density),alpha=log(Population.density)))+facet_grid(~major_teaching)+labs(y="B(6w) - B(3w)",x="Average mobility proportion in 3w-6w",title="B1(6w)-B0(3w) in Only Micropolitan Counties",size="Log of Population Density",color="Majority Teaching Method",alpha="Log of Population Density",fill="Majority Teaching Method")+grid_theme+theme(legend.text = element_text(size=13),legend.title = element_text(size=13),axis.text = element_text(size=11),title=element_text(size=14),strip.text.x = element_text(
        size = 12, face = "bold.italic"
        ))

ggsave("p2.png",width = 10,height = 4)


na.omit(B0B1)%>%
  filter(NCHS.Urban.Rural.Status=="Micropolitan")%>%
  drop_na(major_teaching)%>%
  ggplot(aes(x=avg_full_work_prob,y=max_B1,group=major_teaching,color=major_teaching))+geom_point(aes(size=log(Population.density),alpha=log(Population.density)))+facet_grid(~major_teaching)+labs(y="Max B1",x="Averaged \n%6hr+ Away Home in Fall semester",title="Max B1 in Fall v.s. Average Mobility \nOnly Micropolitan Counties",color="Majority Teaching Method",size = "Log of Population",alpha= "Log of Population" )+grid_theme+theme(legend.text = element_text(size=13),legend.title = element_text(size=13),axis.text = element_text(size=11),title=element_text(size=14),strip.text.x = element_text(
        size = 12, face = "bold.italic"
        ))
ggsave("p3.png",width = 10,height = 4)
```

## Confounders

```{r}
library(ggpubr)
library(PMCMRplus)
require(DescTools)

teaching_profile <- ohio_profile%>%
  inner_join(wide_teaching_enroll,by=c("County"="county"))
teaching_profile$major_teaching <- factor(teaching_profile$major_teaching,levels = c("On Premises","Hybrid","Online Only"))

profile_major_teaching.aov <- aov(Percent.uninsured ~ major_teaching,data = teaching_profile)
summary(profile_major_teaching.aov)

# p-value of .012
stat.test <- PostHocTest(profile_major_teaching.aov, method = "duncan")$major_teaching %>%
  as.data.frame()%>%
  rownames_to_column("group") %>%
  separate(group,"-", into = c("group1","group2")) %>%
  mutate(pval = round(pval,3),
         p = case_when(pval <= .01~ "**",
                       pval <= .05 ~ "*",
                       TRUE ~ "NS"))%>%
  select(group1, group2, pval, p)



teaching_profile%>%
  ggplot(aes(x=major_teaching,y=Percent.uninsured))+geom_boxplot(aes(fill=major_teaching),width=0.6)+
  stat_compare_means(method = "anova",size=6,label.y.npc=0.85)+ 
  stat_pvalue_manual(stat.test, label = "p",y.position = 1, step.increase = 0.15,size = 6,bracket.nudge.y = 8)+
  labs(title="Percent of Uninsured Population by Teaching Method",x="Majority Teaching Method",y="Percent of uninsured population",caption = "Post-hoc pairwise testing with Duncan Method")+
  team_theme+theme(legend.position = "")

ggsave("p3.png",width = 6.5,height = 6)
```



```{r}

senior_major_teaching.aov <- aov(Percent.Population.65..yrs~ major_teaching,data = teaching_profile)

summary(senior_major_teaching.aov)

# p-value of .012
stat.test <- PostHocTest(senior_major_teaching.aov, method = "duncan")$major_teaching %>%
  as.data.frame()%>%
  rownames_to_column("group") %>%
  separate(group,"-", into = c("group1","group2")) %>%
  mutate(pval = round(pval,3),
         p = case_when(pval <= .01~ "**",
                       pval <= .05 ~ "*",
                       TRUE ~ "NS"))%>%
  select(group1, group2, pval, p)

teaching_profile%>%
  ggplot(aes(x=major_teaching,y=Percent.Population.65..yrs))+geom_boxplot(aes(fill=major_teaching),width=0.6)+
  stat_compare_means(method = "anova",size=6,label.y.npc=0.95)+ 
  stat_pvalue_manual(stat.test, label = "p",y.position = 1, step.increase = 0.15,size = 5,bracket.nudge.y = 2)+
  labs(title="Percent of 65years+ Population by Teaching Method",x="Major Teaching Method",y="Percent of 65+yrs population",fill="Majority Teaching Method",caption = "Post-hoc pairwise testing with Duncan Method")+team_theme+theme(legend.position = "")

ggsave("p3.png",width = 6.5,height = 6)
```

```{r}
B_diff_micro%>%
  ggplot(aes(x=log(Population.density),y=acc,
             group=major_teaching,color=major_teaching))+
  geom_point(aes(size=avg2_full_work_prob,alpha=avg2_full_work_prob))+
  geom_smooth(method = "lm", se=F,
              formula = y ~ x,alpha=0.1)+
  theme_minimal()+
  team_theme+
  labs(y="Acceleration",x="Log of Population",
       title="Difference in slope of B(0w) and B(3w)\nOnly Micropolitan Counties",
       color="Majority Teaching Method",
       size = "3w~6w\nAveraged %6hr+ Away Home",
       alpha= "3w~6w\nAveraged %6hr+ Away Home" )
#ggsave("acc.png", width = 7, height = 5)
```
